@@ -1,6 +1,6 @@
 {
   "name": "parse-server",
-  "version": "2.4.2",
+  "version": "2.6.0",
   "description": "An express module providing a Parse-compatible API server",
   "main": "lib/index.js",
   "repository": {
@@ -20,45 +20,45 @@
   "dependencies": {
     "bcryptjs": "2.4.3",
     "body-parser": "1.17.2",
-    "commander": "2.9.0",
+    "commander": "2.11.0",
     "deepcopy": "0.6.3",
-    "express": "4.15.2",
+    "express": "4.15.3",
     "intersect": "1.0.1",
     "lodash": "4.17.4",
-    "lru-cache": "4.0.2",
-    "mime": "1.3.6",
-    "mongodb": "2.2.27",
+    "lru-cache": "4.1.1",
+    "mime": "1.4.0",
+    "mongodb": "2.2.31",
     "multer": "1.3.0",
-    "parse": "1.9.2",
+    "parse": "1.10.0",
     "parse-server-fs-adapter": "1.0.1",
-    "parse-server-push-adapter": "1.3.0",
-    "parse-server-s3-adapter": "1.0.6",
+    "parse-server-push-adapter": "2.0.0",
+    "parse-server-s3-adapter": "1.2.0",
     "parse-server-simple-mailgun-adapter": "1.0.0",
-    "pg-promise": "5.8.1",
-    "redis": "2.7.1",
+    "pg-promise": "6.3.1",
+    "redis": "2.8.0",
     "request": "2.81.0",
-    "semver": "5.3.0",
+    "semver": "5.4.0",
     "tv4": "1.3.0",
     "winston": "2.3.1",
     "winston-daily-rotate-file": "1.4.6",
-    "ws": "3.0.0"
+    "ws": "3.1.0"
   },
   "devDependencies": {
     "babel-cli": "6.24.1",
-    "babel-core": "6.24.1",
+    "babel-core": "6.26.0",
     "babel-eslint": "^7.1.1",
     "babel-plugin-syntax-flow": "6.18.0",
     "babel-plugin-transform-flow-strip-types": "6.22.0",
     "babel-preset-es2015": "6.24.1",
     "babel-preset-stage-0": "6.24.1",
-    "babel-register": "6.24.1",
+    "babel-register": "6.26.0",
     "bcrypt-nodejs": "0.0.3",
-    "cross-env": "5.0.0",
+    "cross-env": "5.0.2",
     "deep-diff": "0.3.8",
-    "eslint": "^3.16.1",
+    "eslint": "^4.4.1",
     "eslint-plugin-flowtype": "^2.25.0",
     "gaze": "1.1.2",
-    "jasmine": "2.6.0",
+    "jasmine": "2.8.0",
     "jasmine-spec-reporter": "^4.1.0",
     "mongodb-runner": "3.5.0",
     "nodemon": "1.11.0",
@@ -84,7 +84,7 @@
     "parse-server": "./bin/parse-server"
   },
   "optionalDependencies": {
-    "bcrypt": "1.0.2",
-    "uws": "^0.14.5"
+    "bcrypt": "1.0.3",
+    "uws": "^8.14.1"
   }
 }
