@@ -5,14 +5,14 @@ const Config = require("../src/Config");
 var loginWithWrongCredentialsShouldFail = function(username, password) {
   return new Promise((resolve, reject) => {
     Parse.User.logIn(username, password)
-    .then(() => reject('login should have failed'))
-    .catch(err => {
-      if (err.message === 'Invalid username/password.') {
-        resolve();
-      } else {
-        reject(err);
-      }
-    });
+      .then(() => reject('login should have failed'))
+      .catch(err => {
+        if (err.message === 'Invalid username/password.') {
+          resolve();
+        } else {
+          reject(err);
+        }
+      });
   });
 };
 
@@ -20,14 +20,14 @@ var isAccountLockoutError = function(username, password, duration, waitTime) {
   return new Promise((resolve, reject) => {
     setTimeout(() => {
       Parse.User.logIn(username, password)
-      .then(() => reject('login should have failed'))
-      .catch(err => {
-        if (err.message === 'Your account is locked due to multiple failed login attempts. Please try again after ' + duration + ' minute(s)') {
-          resolve();
-        } else {
-          reject(err);
-        }
-      });
+        .then(() => reject('login should have failed'))
+        .catch(err => {
+          if (err.message === 'Your account is locked due to multiple failed login attempts. Please try again after ' + duration + ' minute(s)') {
+            resolve();
+          } else {
+            reject(err);
+          }
+        });
     }, waitTime);
   });
 };
@@ -39,26 +39,26 @@ describe("Account Lockout Policy: ", () => {
       appName: 'unlimited',
       publicServerURL: 'http://localhost:1337/1',
     })
-    .then(() => {
-      var user = new Parse.User();
-      user.setUsername('username1');
-      user.setPassword('password');
-      return user.signUp(null);
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username1', 'incorrect password 1');
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username1', 'incorrect password 2');
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username1', 'incorrect password 3');
-    })
-    .then(() => done())
-    .catch(err => {
-      fail('allow unlimited failed login attempts failed: ' + JSON.stringify(err));
-      done();
-    });
+      .then(() => {
+        var user = new Parse.User();
+        user.setUsername('username1');
+        user.setPassword('password');
+        return user.signUp(null);
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username1', 'incorrect password 1');
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username1', 'incorrect password 2');
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username1', 'incorrect password 3');
+      })
+      .then(() => done())
+      .catch(err => {
+        fail('allow unlimited failed login attempts failed: ' + JSON.stringify(err));
+        done();
+      });
   });
 
   it('throw error if duration is set to an invalid number', done => {
@@ -70,19 +70,19 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      new Config('test');
-      fail('set duration to an invalid number test failed');
-      done();
-    })
-    .catch(err => {
-      if (err && err === 'Account lockout duration should be greater than 0 and less than 100000') {
+      .then(() => {
+        new Config('test');
+        fail('set duration to an invalid number test failed');
         done();
-      } else {
-        fail('set duration to an invalid number test failed: ' + JSON.stringify(err));
-        done();
-      }
-    });
+      })
+      .catch(err => {
+        if (err && err === 'Account lockout duration should be greater than 0 and less than 100000') {
+          done();
+        } else {
+          fail('set duration to an invalid number test failed: ' + JSON.stringify(err));
+          done();
+        }
+      });
   });
 
   it('throw error if threshold is set to an invalid number', done => {
@@ -94,19 +94,19 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      new Config('test');
-      fail('set threshold to an invalid number test failed');
-      done();
-    })
-    .catch(err => {
-      if (err && err === 'Account lockout threshold should be an integer greater than 0 and less than 1000') {
+      .then(() => {
+        new Config('test');
+        fail('set threshold to an invalid number test failed');
         done();
-      } else {
-        fail('set threshold to an invalid number test failed: ' + JSON.stringify(err));
-        done();
-      }
-    });
+      })
+      .catch(err => {
+        if (err && err === 'Account lockout threshold should be an integer greater than 0 and less than 1000') {
+          done();
+        } else {
+          fail('set threshold to an invalid number test failed: ' + JSON.stringify(err));
+          done();
+        }
+      });
   });
 
   it('throw error if threshold is < 1', done => {
@@ -118,19 +118,19 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      new Config('test');
-      fail('threshold value < 1 is invalid test failed');
-      done();
-    })
-    .catch(err => {
-      if (err && err === 'Account lockout threshold should be an integer greater than 0 and less than 1000') {
+      .then(() => {
+        new Config('test');
+        fail('threshold value < 1 is invalid test failed');
         done();
-      } else {
-        fail('threshold value < 1 is invalid test failed: ' + JSON.stringify(err));
-        done();
-      }
-    });
+      })
+      .catch(err => {
+        if (err && err === 'Account lockout threshold should be an integer greater than 0 and less than 1000') {
+          done();
+        } else {
+          fail('threshold value < 1 is invalid test failed: ' + JSON.stringify(err));
+          done();
+        }
+      });
   });
 
   it('throw error if threshold is > 999', done => {
@@ -142,19 +142,19 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      new Config('test');
-      fail('threshold value > 999 is invalid test failed');
-      done();
-    })
-    .catch(err => {
-      if (err && err === 'Account lockout threshold should be an integer greater than 0 and less than 1000') {
+      .then(() => {
+        new Config('test');
+        fail('threshold value > 999 is invalid test failed');
         done();
-      } else {
-        fail('threshold value > 999 is invalid test failed: ' + JSON.stringify(err));
-        done();
-      }
-    });
+      })
+      .catch(err => {
+        if (err && err === 'Account lockout threshold should be an integer greater than 0 and less than 1000') {
+          done();
+        } else {
+          fail('threshold value > 999 is invalid test failed: ' + JSON.stringify(err));
+          done();
+        }
+      });
   });
 
   it('throw error if duration is <= 0', done => {
@@ -166,19 +166,19 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      new Config('test');
-      fail('duration value < 1 is invalid test failed');
-      done();
-    })
-    .catch(err => {
-      if (err && err === 'Account lockout duration should be greater than 0 and less than 100000') {
+      .then(() => {
+        new Config('test');
+        fail('duration value < 1 is invalid test failed');
         done();
-      } else {
-        fail('duration value < 1 is invalid test failed: ' + JSON.stringify(err));
-        done();
-      }
-    });
+      })
+      .catch(err => {
+        if (err && err === 'Account lockout duration should be greater than 0 and less than 100000') {
+          done();
+        } else {
+          fail('duration value < 1 is invalid test failed: ' + JSON.stringify(err));
+          done();
+        }
+      });
   });
 
   it('throw error if duration is > 99999', done => {
@@ -190,19 +190,19 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      new Config('test');
-      fail('duration value > 99999 is invalid test failed');
-      done();
-    })
-    .catch(err => {
-      if (err && err === 'Account lockout duration should be greater than 0 and less than 100000') {
+      .then(() => {
+        new Config('test');
+        fail('duration value > 99999 is invalid test failed');
         done();
-      } else {
-        fail('duration value > 99999 is invalid test failed: ' + JSON.stringify(err));
-        done();
-      }
-    });
+      })
+      .catch(err => {
+        if (err && err === 'Account lockout duration should be greater than 0 and less than 100000') {
+          done();
+        } else {
+          fail('duration value > 99999 is invalid test failed: ' + JSON.stringify(err));
+          done();
+        }
+      });
   });
 
   it('lock account if failed login attempts are above threshold', done => {
@@ -214,28 +214,28 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      var user = new Parse.User();
-      user.setUsername("username2");
-      user.setPassword("failedLoginAttemptsThreshold");
-      return user.signUp();
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username2', 'wrong password');
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username2', 'wrong password');
-    })
-    .then(() => {
-      return isAccountLockoutError('username2', 'wrong password', 1, 1);
-    })
-    .then(() => {
-      done();
-    })
-    .catch(err => {
-      fail('lock account after failed login attempts test failed: ' + JSON.stringify(err));
-      done();
-    });
+      .then(() => {
+        var user = new Parse.User();
+        user.setUsername("username2");
+        user.setPassword("failedLoginAttemptsThreshold");
+        return user.signUp();
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username2', 'wrong password');
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username2', 'wrong password');
+      })
+      .then(() => {
+        return isAccountLockoutError('username2', 'wrong password', 1, 1);
+      })
+      .then(() => {
+        done();
+      })
+      .catch(err => {
+        fail('lock account after failed login attempts test failed: ' + JSON.stringify(err));
+        done();
+      });
   });
 
   it('lock account for accountPolicy.duration minutes if failed login attempts are above threshold', done => {
@@ -247,32 +247,32 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      var user = new Parse.User();
-      user.setUsername("username3");
-      user.setPassword("failedLoginAttemptsThreshold");
-      return user.signUp();
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username3', 'wrong password');
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username3', 'wrong password');
-    })
-    .then(() => {
-      return isAccountLockoutError('username3', 'wrong password', 0.05, 1);
-    })
-    .then(() => {
+      .then(() => {
+        var user = new Parse.User();
+        user.setUsername("username3");
+        user.setPassword("failedLoginAttemptsThreshold");
+        return user.signUp();
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username3', 'wrong password');
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username3', 'wrong password');
+      })
+      .then(() => {
+        return isAccountLockoutError('username3', 'wrong password', 0.05, 1);
+      })
+      .then(() => {
       // account should still be locked even after 2 seconds.
-      return isAccountLockoutError('username3', 'wrong password', 0.05, 2000);
-    })
-    .then(() => {
-      done();
-    })
-    .catch(err => {
-      fail('account should be locked for duration mins test failed: ' + JSON.stringify(err));
-      done();
-    });
+        return isAccountLockoutError('username3', 'wrong password', 0.05, 2000);
+      })
+      .then(() => {
+        done();
+      })
+      .catch(err => {
+        fail('account should be locked for duration mins test failed: ' + JSON.stringify(err));
+        done();
+      });
   });
 
   it('allow login for locked account after accountPolicy.duration minutes', done => {
@@ -284,35 +284,35 @@ describe("Account Lockout Policy: ", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      var user = new Parse.User();
-      user.setUsername("username4");
-      user.setPassword("correct password");
-      return user.signUp();
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username4', 'wrong password');
-    })
-    .then(() => {
-      return loginWithWrongCredentialsShouldFail('username4', 'wrong password');
-    })
-    .then(() => {
+      .then(() => {
+        var user = new Parse.User();
+        user.setUsername("username4");
+        user.setPassword("correct password");
+        return user.signUp();
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username4', 'wrong password');
+      })
+      .then(() => {
+        return loginWithWrongCredentialsShouldFail('username4', 'wrong password');
+      })
+      .then(() => {
       // allow locked user to login after 3 seconds with a valid userid and password
-      return new Promise((resolve, reject) => {
-        setTimeout(() => {
-          Parse.User.logIn('username4', 'correct password')
-          .then(() => resolve())
-          .catch(err => reject(err));
-        }, 3001);
+        return new Promise((resolve, reject) => {
+          setTimeout(() => {
+            Parse.User.logIn('username4', 'correct password')
+              .then(() => resolve())
+              .catch(err => reject(err));
+          }, 3001);
+        });
+      })
+      .then(() => {
+        done();
+      })
+      .catch(err => {
+        fail('allow login for locked account after accountPolicy.duration minutes test failed: ' + JSON.stringify(err));
+        done();
       });
-    })
-    .then(() => {
-      done();
-    })
-    .catch(err => {
-      fail('allow login for locked account after accountPolicy.duration minutes test failed: ' + JSON.stringify(err));
-      done();
-    });
   });
 
 })
