@@ -23,28 +23,28 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 0.5, // 0.5 second
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
       // wait for 1 second - simulate user behavior to some extent
-      setTimeout(() => {
-        expect(sendEmailOptions).not.toBeUndefined();
+        setTimeout(() => {
+          expect(sendEmailOptions).not.toBeUndefined();
 
-        request.get(sendEmailOptions.link, {
-          followRedirect: false,
-        }, (error, response) => {
-          expect(response.statusCode).toEqual(302);
-          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=testEmailVerifyTokenValidity&appId=test');
-          done();
-        });
-      }, 1000);
-    }).catch((err) => {
-      jfail(err);
-      done();
-    });
+          request.get(sendEmailOptions.link, {
+            followRedirect: false,
+          }, (error, response) => {
+            expect(response.statusCode).toEqual(302);
+            expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=testEmailVerifyTokenValidity&appId=test');
+            done();
+          });
+        }, 1000);
+      }).catch((err) => {
+        jfail(err);
+        done();
+      });
   });
 
   it('emailVerified should set to false, if the user does not verify their email before the email verify token expires', done => {
@@ -64,35 +64,35 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 0.5, // 0.5 second
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
       // wait for 1 second - simulate user behavior to some extent
-      setTimeout(() => {
-        expect(sendEmailOptions).not.toBeUndefined();
+        setTimeout(() => {
+          expect(sendEmailOptions).not.toBeUndefined();
 
-        request.get(sendEmailOptions.link, {
-          followRedirect: false,
-        }, (error, response) => {
-          expect(response.statusCode).toEqual(302);
-          user.fetch()
-          .then(() => {
-            expect(user.get('emailVerified')).toEqual(false);
-            done();
-          })
-          .catch(() => {
-            jfail(error);
-            done();
+          request.get(sendEmailOptions.link, {
+            followRedirect: false,
+          }, (error, response) => {
+            expect(response.statusCode).toEqual(302);
+            user.fetch()
+              .then(() => {
+                expect(user.get('emailVerified')).toEqual(false);
+                done();
+              })
+              .catch(() => {
+                jfail(error);
+                done();
+              });
           });
-        });
-      }, 1000);
-    }).catch((error) => {
-      jfail(error);
-      done();
-    });
+        }, 1000);
+      }).catch((error) => {
+        jfail(error);
+        done();
+      });
   });
 
   it('if user clicks on the email verify link before email verification token expiration then show the verify email success page', done => {
@@ -112,23 +112,23 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=testEmailVerifyTokenValidity');
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=testEmailVerifyTokenValidity');
+          done();
+        });
+      }).catch((error) => {
+        jfail(error);
         done();
       });
-    }).catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
   it('if user clicks on the email verify link before email verification token expiration then emailVerified should be true', done => {
@@ -148,30 +148,30 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        user.fetch()
-        .then(() => {
-          expect(user.get('emailVerified')).toEqual(true);
-          done();
-        })
-        .catch((error) => {
-          jfail(error);
-          done();
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          user.fetch()
+            .then(() => {
+              expect(user.get('emailVerified')).toEqual(true);
+              done();
+            })
+            .catch((error) => {
+              jfail(error);
+              done();
+            });
         });
+      }).catch((error) => {
+        jfail(error);
+        done();
       });
-    }).catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
   it('if user clicks on the email verify link before email verification token expiration then user should be able to login', done => {
@@ -191,31 +191,31 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        Parse.User.logIn("testEmailVerifyTokenValidity", "expiringToken")
-        .then(user => {
-          expect(typeof user).toBe('object');
-          expect(user.get('emailVerified')).toBe(true);
-          done();
-        })
-        .catch((error) => {
-          jfail(error);
-          done();
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          Parse.User.logIn("testEmailVerifyTokenValidity", "expiringToken")
+            .then(user => {
+              expect(typeof user).toBe('object');
+              expect(user.get('emailVerified')).toBe(true);
+              done();
+            })
+            .catch((error) => {
+              jfail(error);
+              done();
+            });
         });
+      }).catch((error) => {
+        jfail(error);
+        done();
       });
-    }).catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
   it('sets the _email_verify_token_expires_at and _email_verify_token fields after user SignUp', done => {
@@ -235,30 +235,30 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: 'http://localhost:8378/1'
     })
-    .then(() => {
-      user.setUsername('sets_email_verify_token_expires_at');
-      user.setPassword('expiringToken');
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-      const config = new Config('test');
-      return config.database.find('_User', {username: 'sets_email_verify_token_expires_at'});
-    })
-    .then(results => {
-      expect(results.length).toBe(1);
-      const user = results[0];
-      expect(typeof user).toBe('object');
-      expect(user.emailVerified).toEqual(false);
-      expect(typeof user._email_verify_token).toBe('string');
-      expect(typeof user._email_verify_token_expires_at).toBe('object');
-      expect(sendEmailOptions).toBeDefined();
-      done();
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => {
+        user.setUsername('sets_email_verify_token_expires_at');
+        user.setPassword('expiringToken');
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      })
+      .then(() => {
+        const config = new Config('test');
+        return config.database.find('_User', {username: 'sets_email_verify_token_expires_at'});
+      })
+      .then(results => {
+        expect(results.length).toBe(1);
+        const user = results[0];
+        expect(typeof user).toBe('object');
+        expect(user.emailVerified).toEqual(false);
+        expect(typeof user._email_verify_token).toBe('string');
+        expect(typeof user._email_verify_token_expires_at).toBe('object');
+        expect(sendEmailOptions).toBeDefined();
+        done();
+      })
+      .catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('unsets the _email_verify_token_expires_at and _email_verify_token fields in the User class if email verification is successful', done => {
@@ -278,39 +278,39 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("unsets_email_verify_token_expires_at");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        const config = new Config('test');
-        return config.database.find('_User', {username: 'unsets_email_verify_token_expires_at'}).then((results) => {
-          expect(results.length).toBe(1);
-          return results[0];
-        })
-        .then(user => {
-          expect(typeof user).toBe('object');
-          expect(user.emailVerified).toEqual(true);
-          expect(typeof user._email_verify_token).toBe('undefined');
-          expect(typeof user._email_verify_token_expires_at).toBe('undefined');
-          done();
-        })
-        .catch(error => {
-          jfail(error);
-          done();
+      .then(() => {
+        user.setUsername("unsets_email_verify_token_expires_at");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      })
+      .then(() => {
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          const config = new Config('test');
+          return config.database.find('_User', {username: 'unsets_email_verify_token_expires_at'}).then((results) => {
+            expect(results.length).toBe(1);
+            return results[0];
+          })
+            .then(user => {
+              expect(typeof user).toBe('object');
+              expect(user.emailVerified).toEqual(true);
+              expect(typeof user._email_verify_token).toBe('undefined');
+              expect(typeof user._email_verify_token_expires_at).toBe('undefined');
+              done();
+            })
+            .catch(error => {
+              jfail(error);
+              done();
+            });
         });
+      })
+      .catch(error => {
+        jfail(error);
+        done();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('clicking on the email verify link by an email VERIFIED user that was setup before enabling the expire email verify token should show email verify email success', done => {
@@ -332,41 +332,41 @@ describe("Email Verification Token Expiration: ", () => {
 
     // setup server WITHOUT enabling the expire email verify token flag
     reconfigureServer(serverConfig)
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-      return new Promise((resolve, reject) => {
-        request.get(sendEmailOptions.link, { followRedirect: false, })
-        .on('error', error => reject(error))
-        .on('response', (response) => {
-          expect(response.statusCode).toEqual(302);
-          resolve(user.fetch());
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      })
+      .then(() => {
+        return new Promise((resolve, reject) => {
+          request.get(sendEmailOptions.link, { followRedirect: false, })
+            .on('error', error => reject(error))
+            .on('response', (response) => {
+              expect(response.statusCode).toEqual(302);
+              resolve(user.fetch());
+            });
         });
-      });
-    })
-    .then(() => {
-      expect(user.get('emailVerified')).toEqual(true);
-      // RECONFIGURE the server i.e., ENABLE the expire email verify token flag
-      serverConfig.emailVerifyTokenValidityDuration = 5; // 5 seconds
-      return reconfigureServer(serverConfig);
-    })
-    .then(() => {
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=testEmailVerifyTokenValidity');
+      })
+      .then(() => {
+        expect(user.get('emailVerified')).toEqual(true);
+        // RECONFIGURE the server i.e., ENABLE the expire email verify token flag
+        serverConfig.emailVerifyTokenValidityDuration = 5; // 5 seconds
+        return reconfigureServer(serverConfig);
+      })
+      .then(() => {
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=testEmailVerifyTokenValidity');
+          done();
+        });
+      })
+      .catch((error) => {
+        jfail(error);
         done();
       });
-    })
-    .catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
   it('clicking on the email verify link by an email UNVERIFIED user that was setup before enabling the expire email verify token should show invalid verficiation link page', done => {
@@ -388,35 +388,35 @@ describe("Email Verification Token Expiration: ", () => {
 
     // setup server WITHOUT enabling the expire email verify token flag
     reconfigureServer(serverConfig)
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
+      .then(() => {
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      })
+      .then(() => {
       // just get the user again - DO NOT email verify the user
-      return user.fetch();
-    })
-    .then(() => {
-      expect(user.get('emailVerified')).toEqual(false);
-      // RECONFIGURE the server i.e., ENABLE the expire email verify token flag
-      serverConfig.emailVerifyTokenValidityDuration = 5; // 5 seconds
-      return reconfigureServer(serverConfig);
-    })
-    .then(() => {
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=testEmailVerifyTokenValidity&appId=test');
+        return user.fetch();
+      })
+      .then(() => {
+        expect(user.get('emailVerified')).toEqual(false);
+        // RECONFIGURE the server i.e., ENABLE the expire email verify token flag
+        serverConfig.emailVerifyTokenValidityDuration = 5; // 5 seconds
+        return reconfigureServer(serverConfig);
+      })
+      .then(() => {
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=testEmailVerifyTokenValidity&appId=test');
+          done();
+        });
+      })
+      .catch((error) => {
+        jfail(error);
         done();
       });
-    })
-    .catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
   it('setting the email on the user should set a new email verification token and new expiration date for the token when expire email verify token flag is set', done => {
@@ -441,46 +441,46 @@ describe("Email Verification Token Expiration: ", () => {
     };
 
     reconfigureServer(serverConfig)
-    .then(() => {
-      user.setUsername("newEmailVerifyTokenOnEmailReset");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-      const config = new Config('test');
-      return config.database.find('_User', {username: 'newEmailVerifyTokenOnEmailReset'}).then((results) => {
-        return results[0];
-      });
-    })
-    .then(userFromDb => {
-      expect(typeof userFromDb).toBe('object');
-      userBeforeEmailReset = userFromDb;
+      .then(() => {
+        user.setUsername("newEmailVerifyTokenOnEmailReset");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      })
+      .then(() => {
+        const config = new Config('test');
+        return config.database.find('_User', {username: 'newEmailVerifyTokenOnEmailReset'}).then((results) => {
+          return results[0];
+        });
+      })
+      .then(userFromDb => {
+        expect(typeof userFromDb).toBe('object');
+        userBeforeEmailReset = userFromDb;
 
-      // trigger another token generation by setting the email
-      user.set('email', 'user@parse.com');
-      return new Promise((resolve) => {
+        // trigger another token generation by setting the email
+        user.set('email', 'user@parse.com');
+        return new Promise((resolve) => {
         // wait for half a sec to get a new expiration time
-        setTimeout(() => resolve(user.save()), 500);
+          setTimeout(() => resolve(user.save()), 500);
+        });
+      })
+      .then(() => {
+        const config = new Config('test');
+        return config.database.find('_User', {username: 'newEmailVerifyTokenOnEmailReset'}).then((results) => {
+          return results[0];
+        });
+      })
+      .then(userAfterEmailReset => {
+        expect(typeof userAfterEmailReset).toBe('object');
+        expect(userBeforeEmailReset._email_verify_token).not.toEqual(userAfterEmailReset._email_verify_token);
+        expect(userBeforeEmailReset._email_verify_token_expires_at).not.toEqual(userAfterEmailReset.__email_verify_token_expires_at);
+        expect(sendEmailOptions).toBeDefined();
+        done();
+      })
+      .catch((error) => {
+        jfail(error);
+        done();
       });
-    })
-    .then(() => {
-      const config = new Config('test');
-      return config.database.find('_User', {username: 'newEmailVerifyTokenOnEmailReset'}).then((results) => {
-        return results[0];
-      });
-    })
-    .then(userAfterEmailReset => {
-      expect(typeof userAfterEmailReset).toBe('object');
-      expect(userBeforeEmailReset._email_verify_token).not.toEqual(userAfterEmailReset._email_verify_token);
-      expect(userBeforeEmailReset._email_verify_token_expires_at).not.toEqual(userAfterEmailReset.__email_verify_token_expires_at);
-      expect(sendEmailOptions).toBeDefined();
-      done();
-    })
-    .catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
   it('should send a new verification email when a resend is requested and the user is UNVERIFIED', done => {
@@ -502,39 +502,39 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: 'http://localhost:8378/1'
     })
-    .then(() => {
-      user.setUsername('resends_verification_token');
-      user.setPassword('expiringToken');
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-      expect(sendVerificationEmailCallCount).toBe(1);
-
-      return requestp.post({
-        uri: 'http://localhost:8378/1/verificationEmailRequest',
-        body: {
-          email: 'user@parse.com'
-        },
-        headers: {
-          'X-Parse-Application-Id': Parse.applicationId,
-          'X-Parse-REST-API-Key': 'rest',
-        },
-        json: true,
-        resolveWithFullResponse: true,
-        simple: false // this promise is only rejected if the call itself failed
+      .then(() => {
+        user.setUsername('resends_verification_token');
+        user.setPassword('expiringToken');
+        user.set('email', 'user@parse.com');
+        return user.signUp();
       })
-      .then((response) => {
-        expect(response.statusCode).toBe(200);
-        expect(sendVerificationEmailCallCount).toBe(2);
-        expect(sendEmailOptions).toBeDefined();
+      .then(() => {
+        expect(sendVerificationEmailCallCount).toBe(1);
+
+        return requestp.post({
+          uri: 'http://localhost:8378/1/verificationEmailRequest',
+          body: {
+            email: 'user@parse.com'
+          },
+          headers: {
+            'X-Parse-Application-Id': Parse.applicationId,
+            'X-Parse-REST-API-Key': 'rest',
+          },
+          json: true,
+          resolveWithFullResponse: true,
+          simple: false // this promise is only rejected if the call itself failed
+        })
+          .then((response) => {
+            expect(response.statusCode).toBe(200);
+            expect(sendVerificationEmailCallCount).toBe(2);
+            expect(sendEmailOptions).toBeDefined();
+            done();
+          });
+      })
+      .catch(error => {
+        jfail(error);
         done();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('should not send a new verification email when a resend is requested and the user is VERIFIED', done => {
@@ -556,49 +556,49 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: 'http://localhost:8378/1'
     })
-    .then(() => {
-      user.setUsername('no_new_verification_token_once_verified');
-      user.setPassword('expiringToken');
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-      return requestp.get({
-        url: sendEmailOptions.link,
-        followRedirect: false,
-        resolveWithFullResponse: true,
-        simple: false
+      .then(() => {
+        user.setUsername('no_new_verification_token_once_verified');
+        user.setPassword('expiringToken');
+        user.set('email', 'user@parse.com');
+        return user.signUp();
       })
-      .then((response) => {
-        expect(response.statusCode).toEqual(302);
-      });
-    })
-    .then(() => {
-      expect(sendVerificationEmailCallCount).toBe(1);
-
-      return requestp.post({
-        uri: 'http://localhost:8378/1/verificationEmailRequest',
-        body: {
-          email: 'user@parse.com'
-        },
-        headers: {
-          'X-Parse-Application-Id': Parse.applicationId,
-          'X-Parse-REST-API-Key': 'rest',
-        },
-        json: true,
-        resolveWithFullResponse: true,
-        simple: false // this promise is only rejected if the call itself failed
+      .then(() => {
+        return requestp.get({
+          url: sendEmailOptions.link,
+          followRedirect: false,
+          resolveWithFullResponse: true,
+          simple: false
+        })
+          .then((response) => {
+            expect(response.statusCode).toEqual(302);
+          });
       })
-      .then((response) => {
-        expect(response.statusCode).toBe(400);
+      .then(() => {
         expect(sendVerificationEmailCallCount).toBe(1);
+
+        return requestp.post({
+          uri: 'http://localhost:8378/1/verificationEmailRequest',
+          body: {
+            email: 'user@parse.com'
+          },
+          headers: {
+            'X-Parse-Application-Id': Parse.applicationId,
+            'X-Parse-REST-API-Key': 'rest',
+          },
+          json: true,
+          resolveWithFullResponse: true,
+          simple: false // this promise is only rejected if the call itself failed
+        })
+          .then((response) => {
+            expect(response.statusCode).toBe(400);
+            expect(sendVerificationEmailCallCount).toBe(1);
+            done();
+          });
+      })
+      .catch(error => {
+        jfail(error);
         done();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('should not send a new verification email if this user does not exist', done => {
@@ -619,31 +619,31 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: 'http://localhost:8378/1'
     })
-    .then(() => {
-      return requestp.post({
-        uri: 'http://localhost:8378/1/verificationEmailRequest',
-        body: {
-          email: 'user@parse.com'
-        },
-        headers: {
-          'X-Parse-Application-Id': Parse.applicationId,
-          'X-Parse-REST-API-Key': 'rest',
-        },
-        json: true,
-        resolveWithFullResponse: true,
-        simple: false
+      .then(() => {
+        return requestp.post({
+          uri: 'http://localhost:8378/1/verificationEmailRequest',
+          body: {
+            email: 'user@parse.com'
+          },
+          headers: {
+            'X-Parse-Application-Id': Parse.applicationId,
+            'X-Parse-REST-API-Key': 'rest',
+          },
+          json: true,
+          resolveWithFullResponse: true,
+          simple: false
+        })
+          .then(response => {
+            expect(response.statusCode).toBe(400);
+            expect(sendVerificationEmailCallCount).toBe(0);
+            expect(sendEmailOptions).not.toBeDefined();
+            done();
+          });
       })
-      .then(response => {
-        expect(response.statusCode).toBe(400);
-        expect(sendVerificationEmailCallCount).toBe(0);
-        expect(sendEmailOptions).not.toBeDefined();
+      .catch(error => {
+        jfail(error);
         done();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('should fail if no email is supplied', done => {
@@ -664,30 +664,30 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: 'http://localhost:8378/1'
     })
-    .then(() => {
-      request.post({
-        uri: 'http://localhost:8378/1/verificationEmailRequest',
-        body: {},
-        headers: {
-          'X-Parse-Application-Id': Parse.applicationId,
-          'X-Parse-REST-API-Key': 'rest',
-        },
-        json: true,
-        resolveWithFullResponse: true,
-        simple: false
-      }, (err, response) => {
-        expect(response.statusCode).toBe(400);
-        expect(response.body.code).toBe(Parse.Error.EMAIL_MISSING);
-        expect(response.body.error).toBe('you must provide an email');
-        expect(sendVerificationEmailCallCount).toBe(0);
-        expect(sendEmailOptions).not.toBeDefined();
+      .then(() => {
+        request.post({
+          uri: 'http://localhost:8378/1/verificationEmailRequest',
+          body: {},
+          headers: {
+            'X-Parse-Application-Id': Parse.applicationId,
+            'X-Parse-REST-API-Key': 'rest',
+          },
+          json: true,
+          resolveWithFullResponse: true,
+          simple: false
+        }, (err, response) => {
+          expect(response.statusCode).toBe(400);
+          expect(response.body.code).toBe(Parse.Error.EMAIL_MISSING);
+          expect(response.body.error).toBe('you must provide an email');
+          expect(sendVerificationEmailCallCount).toBe(0);
+          expect(sendEmailOptions).not.toBeDefined();
+          done();
+        });
+      })
+      .catch(error => {
+        jfail(error);
         done();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('should fail if email is not a string', done => {
@@ -708,30 +708,30 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: 'http://localhost:8378/1'
     })
-    .then(() => {
-      request.post({
-        uri: 'http://localhost:8378/1/verificationEmailRequest',
-        body: {email: 3},
-        headers: {
-          'X-Parse-Application-Id': Parse.applicationId,
-          'X-Parse-REST-API-Key': 'rest',
-        },
-        json: true,
-        resolveWithFullResponse: true,
-        simple: false
-      }, (err, response) => {
-        expect(response.statusCode).toBe(400);
-        expect(response.body.code).toBe(Parse.Error.INVALID_EMAIL_ADDRESS);
-        expect(response.body.error).toBe('you must provide a valid email string');
-        expect(sendVerificationEmailCallCount).toBe(0);
-        expect(sendEmailOptions).not.toBeDefined();
+      .then(() => {
+        request.post({
+          uri: 'http://localhost:8378/1/verificationEmailRequest',
+          body: {email: 3},
+          headers: {
+            'X-Parse-Application-Id': Parse.applicationId,
+            'X-Parse-REST-API-Key': 'rest',
+          },
+          json: true,
+          resolveWithFullResponse: true,
+          simple: false
+        }, (err, response) => {
+          expect(response.statusCode).toBe(400);
+          expect(response.body.code).toBe(Parse.Error.INVALID_EMAIL_ADDRESS);
+          expect(response.body.error).toBe('you must provide a valid email string');
+          expect(sendVerificationEmailCallCount).toBe(0);
+          expect(sendEmailOptions).not.toBeDefined();
+          done();
+        });
+      })
+      .catch(error => {
+        jfail(error);
         done();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('client should not see the _email_verify_token_expires_at field', done => {
@@ -751,30 +751,30 @@ describe("Email Verification Token Expiration: ", () => {
       emailVerifyTokenValidityDuration: 5, // 5 seconds
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setUsername("testEmailVerifyTokenValidity");
-      user.setPassword("expiringToken");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    })
-    .then(() => {
-
-      user.fetch()
       .then(() => {
-        expect(user.get('emailVerified')).toEqual(false);
-        expect(typeof user.get('_email_verify_token_expires_at')).toBe('undefined');
-        expect(sendEmailOptions).toBeDefined();
-        done();
+        user.setUsername("testEmailVerifyTokenValidity");
+        user.setPassword("expiringToken");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
       })
-      .catch(error => {
+      .then(() => {
+
+        user.fetch()
+          .then(() => {
+            expect(user.get('emailVerified')).toEqual(false);
+            expect(typeof user.get('_email_verify_token_expires_at')).toBe('undefined');
+            expect(sendEmailOptions).toBeDefined();
+            done();
+          })
+          .catch(error => {
+            jfail(error);
+            done();
+          });
+
+      }).catch((error) => {
         jfail(error);
         done();
       });
-
-    }).catch((error) => {
-      jfail(error);
-      done();
-    });
   });
 
 })
