@@ -28,18 +28,18 @@ describe('InstallationsRouter', () => {
 
     var router = new InstallationsRouter();
     rest.create(config, auth.nobody(config), '_Installation', androidDeviceRequest)
-    .then(() => {
-      return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
-    }).then(() => {
-      return router.handleFind(request);
-    }).then((res) => {
-      var results = res.response.results;
-      expect(results.length).toEqual(1);
-      done();
-    }).catch((err) => {
-      fail(JSON.stringify(err));
-      done();
-    });
+      .then(() => {
+        return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
+      }).then(() => {
+        return router.handleFind(request);
+      }).then((res) => {
+        var results = res.response.results;
+        expect(results.length).toEqual(1);
+        done();
+      }).catch((err) => {
+        fail(JSON.stringify(err));
+        done();
+      });
   });
 
   it('uses find condition from request.query', (done) => {
@@ -66,18 +66,18 @@ describe('InstallationsRouter', () => {
 
     var router = new InstallationsRouter();
     rest.create(config, auth.nobody(config), '_Installation', androidDeviceRequest)
-        .then(() => {
-          return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
-        }).then(() => {
-          return router.handleFind(request);
-        }).then((res) => {
-          var results = res.response.results;
-          expect(results.length).toEqual(1);
-          done();
-        }).catch((err) => {
-          jfail(err);
-          done();
-        });
+      .then(() => {
+        return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
+      }).then(() => {
+        return router.handleFind(request);
+      }).then((res) => {
+        var results = res.response.results;
+        expect(results.length).toEqual(1);
+        done();
+      }).catch((err) => {
+        jfail(err);
+        done();
+      });
   });
 
   it('query installations with limit = 0', (done) => {
@@ -103,18 +103,18 @@ describe('InstallationsRouter', () => {
     new Config('test');
     var router = new InstallationsRouter();
     rest.create(config, auth.nobody(config), '_Installation', androidDeviceRequest)
-        .then(() => {
-          return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
-        }).then(() => {
-          return router.handleFind(request);
-        }).then((res) => {
-          var response = res.response;
-          expect(response.results.length).toEqual(0);
-          done();
-        }).catch((err) => {
-          fail(JSON.stringify(err));
-          done();
-        });
+      .then(() => {
+        return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
+      }).then(() => {
+        return router.handleFind(request);
+      }).then((res) => {
+        var response = res.response;
+        expect(response.results.length).toEqual(0);
+        done();
+      }).catch((err) => {
+        fail(JSON.stringify(err));
+        done();
+      });
   });
 
   it('query installations with count = 1', done => {
@@ -139,18 +139,18 @@ describe('InstallationsRouter', () => {
 
     var router = new InstallationsRouter();
     rest.create(config, auth.nobody(config), '_Installation', androidDeviceRequest)
-    .then(() => rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest))
-    .then(() => router.handleFind(request))
-    .then((res) => {
-      var response = res.response;
-      expect(response.results.length).toEqual(2);
-      expect(response.count).toEqual(2);
-      done();
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    })
+      .then(() => rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest))
+      .then(() => router.handleFind(request))
+      .then((res) => {
+        var response = res.response;
+        expect(response.results.length).toEqual(2);
+        expect(response.count).toEqual(2);
+        done();
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
+        done();
+      })
   });
 
   it('query installations with limit = 0 and count = 1', (done) => {
@@ -176,18 +176,18 @@ describe('InstallationsRouter', () => {
 
     var router = new InstallationsRouter();
     rest.create(config, auth.nobody(config), '_Installation', androidDeviceRequest)
-        .then(() => {
-          return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
-        }).then(() => {
-          return router.handleFind(request);
-        }).then((res) => {
-          var response = res.response;
-          expect(response.results.length).toEqual(0);
-          expect(response.count).toEqual(2);
-          done();
-        }).catch((err) => {
-          fail(JSON.stringify(err));
-          done();
-        });
+      .then(() => {
+        return rest.create(config, auth.nobody(config), '_Installation', iosDeviceRequest);
+      }).then(() => {
+        return router.handleFind(request);
+      }).then((res) => {
+        var response = res.response;
+        expect(response.results.length).toEqual(0);
+        expect(response.count).toEqual(2);
+        done();
+      }).catch((err) => {
+        fail(JSON.stringify(err));
+        done();
+      });
   });
 });
