@@ -45,29 +45,29 @@ describe('Parse.Push', () => {
         adapter: pushAdapter
       }
     })
-    .then(() => {
-      var installations = [];
-      while(installations.length != 10) {
-        var installation = new Parse.Object("_Installation");
-        installation.set("installationId", "installation_" + installations.length);
-        installation.set("deviceToken","device_token_" + installations.length)
-        installation.set("badge", installations.length);
-        installation.set("originalBadge", installations.length);
-        installation.set("deviceType", "ios");
-        installations.push(installation);
-      }
-      return Parse.Object.saveAll(installations);
-    })
-    .then(() => {
-      return {
-        sendToInstallationSpy,
-      };
-    })
-    .catch((err) => {
-      console.error(err);
+      .then(() => {
+        var installations = [];
+        while(installations.length != 10) {
+          var installation = new Parse.Object("_Installation");
+          installation.set("installationId", "installation_" + installations.length);
+          installation.set("deviceToken","device_token_" + installations.length)
+          installation.set("badge", installations.length);
+          installation.set("originalBadge", installations.length);
+          installation.set("deviceType", "ios");
+          installations.push(installation);
+        }
+        return Parse.Object.saveAll(installations);
+      })
+      .then(() => {
+        return {
+          sendToInstallationSpy,
+        };
+      })
+      .catch((err) => {
+        console.error(err);
 
-      throw err;
-    })
+        throw err;
+      })
   }
 
   it('should properly send push', (done) => {
@@ -81,12 +81,12 @@ describe('Parse.Push', () => {
           alert: 'Hello world!'
         }
       }, {useMasterKey: true})
-      .then(() => {
-        return delayPromise(500);
-      })
-      .then(() => {
-        expect(sendToInstallationSpy.calls.count()).toEqual(10);
-      })
+        .then(() => {
+          return delayPromise(500);
+        })
+        .then(() => {
+          expect(sendToInstallationSpy.calls.count()).toEqual(10);
+        })
     }).then(() => {
       done();
     }).catch((err) => {
@@ -118,71 +118,7 @@ describe('Parse.Push', () => {
 
   it('should not allow clients to query _PushStatus', done => {
     setup()
-    .then(() => Parse.Push.send({
-      where: {
-        deviceType: 'ios'
-      },
-      data: {
-        badge: 'increment',
-        alert: 'Hello world!'
-      }
-    }, {useMasterKey: true}))
-    .then(() => {
-      request.get({
-        url: 'http://localhost:8378/1/classes/_PushStatus',
-        json: true,
-        headers: {
-          'X-Parse-Application-Id': 'test',
-        },
-      }, (error, response, body) => {
-        expect(body.error).toEqual('unauthorized');
-        done();
-      });
-    }).catch((err) => {
-      jfail(err);
-      done();
-    });
-  });
-
-  it('should allow master key to query _PushStatus', done => {
-    setup()
-    .then(() => Parse.Push.send({
-      where: {
-        deviceType: 'ios'
-      },
-      data: {
-        badge: 'increment',
-        alert: 'Hello world!'
-      }
-    }, {useMasterKey: true}))
-    .then(() => {
-      request.get({
-        url: 'http://localhost:8378/1/classes/_PushStatus',
-        json: true,
-        headers: {
-          'X-Parse-Application-Id': 'test',
-          'X-Parse-Master-Key': 'test',
-        },
-      }, (error, response, body) => {
-        try {
-          expect(body.results.length).toEqual(1);
-          expect(body.results[0].query).toEqual('{"deviceType":"ios"}');
-          expect(body.results[0].payload).toEqual('{"badge":"increment","alert":"Hello world!"}');
-        } catch(e) {
-          jfail(e);
-        }
-        done();
-      });
-    }).catch((err) => {
-      jfail(err);
-      done();
-    });
-  });
-
-  it('should throw error if missing push configuration', done => {
-    reconfigureServer({push: null})
-    .then(() => {
-      return Parse.Push.send({
+      .then(() => Parse.Push.send({
         where: {
           deviceType: 'ios'
         },
@@ -190,15 +126,79 @@ describe('Parse.Push', () => {
           badge: 'increment',
           alert: 'Hello world!'
         }
-      }, {useMasterKey: true})
-    }).then(() => {
-      fail('should not succeed');
-    }, (err) => {
-      expect(err.code).toEqual(Parse.Error.PUSH_MISCONFIGURED);
-      done();
-    }).catch((err) => {
-      jfail(err);
-      done();
-    });
+      }, {useMasterKey: true}))
+      .then(() => {
+        request.get({
+          url: 'http://localhost:8378/1/classes/_PushStatus',
+          json: true,
+          headers: {
+            'X-Parse-Application-Id': 'test',
+          },
+        }, (error, response, body) => {
+          expect(body.error).toEqual('unauthorized');
+          done();
+        });
+      }).catch((err) => {
+        jfail(err);
+        done();
+      });
+  });
+
+  it('should allow master key to query _PushStatus', done => {
+    setup()
+      .then(() => Parse.Push.send({
+        where: {
+          deviceType: 'ios'
+        },
+        data: {
+          badge: 'increment',
+          alert: 'Hello world!'
+        }
+      }, {useMasterKey: true}))
+      .then(() => {
+        request.get({
+          url: 'http://localhost:8378/1/classes/_PushStatus',
+          json: true,
+          headers: {
+            'X-Parse-Application-Id': 'test',
+            'X-Parse-Master-Key': 'test',
+          },
+        }, (error, response, body) => {
+          try {
+            expect(body.results.length).toEqual(1);
+            expect(body.results[0].query).toEqual('{"deviceType":"ios"}');
+            expect(body.results[0].payload).toEqual('{"badge":"increment","alert":"Hello world!"}');
+          } catch(e) {
+            jfail(e);
+          }
+          done();
+        });
+      }).catch((err) => {
+        jfail(err);
+        done();
+      });
+  });
+
+  it('should throw error if missing push configuration', done => {
+    reconfigureServer({push: null})
+      .then(() => {
+        return Parse.Push.send({
+          where: {
+            deviceType: 'ios'
+          },
+          data: {
+            badge: 'increment',
+            alert: 'Hello world!'
+          }
+        }, {useMasterKey: true})
+      }).then(() => {
+        fail('should not succeed');
+      }, (err) => {
+        expect(err.code).toEqual(Parse.Error.PUSH_MISCONFIGURED);
+        done();
+      }).catch((err) => {
+        jfail(err);
+        done();
+      });
   });
 });
