@@ -58,20 +58,20 @@ describe('Parse.ACL', () => {
             ok(object.get("ACL"));
             // Start making requests by the public, which should all fail.
             Parse.User.logOut()
-            .then(() => {
+              .then(() => {
               // Get
-              var query = new Parse.Query(TestObject);
-              query.get(object.id, {
-                success: function() {
-                  fail('Should not have retrieved the object.');
-                  done();
-                },
-                error: function(model, error) {
-                  equal(error.code, Parse.Error.OBJECT_NOT_FOUND);
-                  done();
-                }
+                var query = new Parse.Query(TestObject);
+                query.get(object.id, {
+                  success: function() {
+                    fail('Should not have retrieved the object.');
+                    done();
+                  },
+                  error: function(model, error) {
+                    equal(error.code, Parse.Error.OBJECT_NOT_FOUND);
+                    done();
+                  }
+                });
               });
-            });
           }
         });
       }
@@ -98,16 +98,16 @@ describe('Parse.ACL', () => {
 
             // Start making requests by the public, which should all fail.
             Parse.User.logOut()
-            .then(() => {
+              .then(() => {
               // Find
-              var query = new Parse.Query(TestObject);
-              query.find({
-                success: function(results) {
-                  equal(results.length, 0);
-                  done();
-                }
+                var query = new Parse.Query(TestObject);
+                query.find({
+                  success: function(results) {
+                    equal(results.length, 0);
+                    done();
+                  }
+                });
               });
-            });
 
           }
         });
@@ -135,19 +135,19 @@ describe('Parse.ACL', () => {
 
             // Start making requests by the public, which should all fail.
             Parse.User.logOut()
-            .then(() => {
+              .then(() => {
               // Update
-              object.set("foo", "bar");
-              object.save(null, {
-                success: function() {
-                  fail('Should not have been able to update the object.');
-                  done();
-                }, error: function(model, err) {
-                  equal(err.code, Parse.Error.OBJECT_NOT_FOUND);
-                  done();
-                }
+                object.set("foo", "bar");
+                object.save(null, {
+                  success: function() {
+                    fail('Should not have been able to update the object.');
+                    done();
+                  }, error: function(model, err) {
+                    equal(err.code, Parse.Error.OBJECT_NOT_FOUND);
+                    done();
+                  }
+                });
               });
-            });
           }
         });
       }
@@ -174,14 +174,14 @@ describe('Parse.ACL', () => {
 
             // Start making requests by the public, which should all fail.
             Parse.User.logOut()
-            .then(() => object.destroy())
-            .then(() => {
-              fail('destroy should fail');
-              done();
-            }, error => {
-              expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
-              done();
-            });
+              .then(() => object.destroy())
+              .then(() => {
+                fail('destroy should fail');
+                done();
+              }, error => {
+                expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
+                done();
+              });
           }
         });
       }
@@ -207,26 +207,26 @@ describe('Parse.ACL', () => {
             ok(object.get("ACL"));
 
             Parse.User.logOut()
-            .then(() => {
-              Parse.User.logIn("alice", "wonderland", {
-                success: function() {
+              .then(() => {
+                Parse.User.logIn("alice", "wonderland", {
+                  success: function() {
                   // Get
-                  var query = new Parse.Query(TestObject);
-                  query.get(object.id, {
-                    success: function(result) {
-                      ok(result);
-                      equal(result.id, object.id);
-                      equal(result.getACL().getReadAccess(user), true);
-                      equal(result.getACL().getWriteAccess(user), true);
-                      equal(result.getACL().getPublicReadAccess(), false);
-                      equal(result.getACL().getPublicWriteAccess(), false);
-                      ok(object.get("ACL"));
-                      done();
-                    }
-                  });
-                }
+                    var query = new Parse.Query(TestObject);
+                    query.get(object.id, {
+                      success: function(result) {
+                        ok(result);
+                        equal(result.id, object.id);
+                        equal(result.getACL().getReadAccess(user), true);
+                        equal(result.getACL().getWriteAccess(user), true);
+                        equal(result.getACL().getPublicReadAccess(), false);
+                        equal(result.getACL().getPublicWriteAccess(), false);
+                        ok(object.get("ACL"));
+                        done();
+                      }
+                    });
+                  }
+                });
               });
-            });
           }
         });
       }
@@ -252,31 +252,31 @@ describe('Parse.ACL', () => {
             ok(object.get("ACL"));
 
             Parse.User.logOut()
-            .then(() => {
-              Parse.User.logIn("alice", "wonderland", {
-                success: function() {
+              .then(() => {
+                Parse.User.logIn("alice", "wonderland", {
+                  success: function() {
                   // Find
-                  var query = new Parse.Query(TestObject);
-                  query.find({
-                    success: function(results) {
-                      equal(results.length, 1);
-                      var result = results[0];
-                      ok(result);
-                      if (!result) {
-                        return fail();
+                    var query = new Parse.Query(TestObject);
+                    query.find({
+                      success: function(results) {
+                        equal(results.length, 1);
+                        var result = results[0];
+                        ok(result);
+                        if (!result) {
+                          return fail();
+                        }
+                        equal(result.id, object.id);
+                        equal(result.getACL().getReadAccess(user), true);
+                        equal(result.getACL().getWriteAccess(user), true);
+                        equal(result.getACL().getPublicReadAccess(), false);
+                        equal(result.getACL().getPublicWriteAccess(), false);
+                        ok(object.get("ACL"));
+                        done();
                       }
-                      equal(result.id, object.id);
-                      equal(result.getACL().getReadAccess(user), true);
-                      equal(result.getACL().getWriteAccess(user), true);
-                      equal(result.getACL().getPublicReadAccess(), false);
-                      equal(result.getACL().getPublicWriteAccess(), false);
-                      ok(object.get("ACL"));
-                      done();
-                    }
-                  });
-                }
+                    });
+                  }
+                });
               });
-            });
           }
         });
       }
@@ -302,19 +302,19 @@ describe('Parse.ACL', () => {
             ok(object.get("ACL"));
 
             Parse.User.logOut()
-            .then(() => {
-              Parse.User.logIn("alice", "wonderland", {
-                success: function() {
+              .then(() => {
+                Parse.User.logIn("alice", "wonderland", {
+                  success: function() {
                   // Update
-                  object.set("foo", "bar");
-                  object.save(null, {
-                    success: function() {
-                      done();
-                    }
-                  });
-                }
+                    object.set("foo", "bar");
+                    object.save(null, {
+                      success: function() {
+                        done();
+                      }
+                    });
+                  }
+                });
               });
-            });
           }
         });
       }
@@ -340,18 +340,18 @@ describe('Parse.ACL', () => {
             ok(object.get("ACL"));
 
             Parse.User.logOut()
-            .then(() => {
-              Parse.User.logIn("alice", "wonderland", {
-                success: function() {
+              .then(() => {
+                Parse.User.logIn("alice", "wonderland", {
+                  success: function() {
                   // Delete
-                  object.destroy({
-                    success: function() {
-                      done();
-                    }
-                  });
-                }
+                    object.destroy({
+                      success: function() {
+                        done();
+                      }
+                    });
+                  }
+                });
               });
-            });
           }
         });
       }
@@ -387,17 +387,17 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Get
-                  var query = new Parse.Query(TestObject);
-                  query.get(object.id, {
-                    success: function(result) {
-                      ok(result);
-                      equal(result.id, object.id);
-                      done();
-                    }
+                    var query = new Parse.Query(TestObject);
+                    query.get(object.id, {
+                      success: function(result) {
+                        ok(result);
+                        equal(result.id, object.id);
+                        done();
+                      }
+                    });
                   });
-                });
               }
             });
           }
@@ -435,19 +435,19 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Find
-                  var query = new Parse.Query(TestObject);
-                  query.find({
-                    success: function(results) {
-                      equal(results.length, 1);
-                      var result = results[0];
-                      ok(result);
-                      equal(result.id, object.id);
-                      done();
-                    }
+                    var query = new Parse.Query(TestObject);
+                    query.find({
+                      success: function(results) {
+                        equal(results.length, 1);
+                        var result = results[0];
+                        ok(result);
+                        equal(result.id, object.id);
+                        done();
+                      }
+                    });
                   });
-                });
               }
             });
           }
@@ -485,16 +485,16 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Update
-                  object.set("foo", "bar");
-                  object.save().then(() => {
-                    fail('the save should fail');
-                  }, error => {
-                    expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
-                    done();
+                    object.set("foo", "bar");
+                    object.save().then(() => {
+                      fail('the save should fail');
+                    }, error => {
+                      expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
+                      done();
+                    });
                   });
-                });
               }
             });
           }
@@ -532,13 +532,13 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => object.destroy())
-                .then(() => {
-                  fail('expected failure');
-                }, error => {
-                  expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
-                  done();
-                });
+                  .then(() => object.destroy())
+                  .then(() => {
+                    fail('expected failure');
+                  }, error => {
+                    expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
+                    done();
+                  });
               }
             });
           }
@@ -576,16 +576,16 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Get
-                  var query = new Parse.Query(TestObject);
-                  query.get(object.id, {
-                    error: function(model, error) {
-                      equal(error.code, Parse.Error.OBJECT_NOT_FOUND);
-                      done();
-                    }
+                    var query = new Parse.Query(TestObject);
+                    query.get(object.id, {
+                      error: function(model, error) {
+                        equal(error.code, Parse.Error.OBJECT_NOT_FOUND);
+                        done();
+                      }
+                    });
                   });
-                });
               }
             });
           }
@@ -623,16 +623,16 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Find
-                  var query = new Parse.Query(TestObject);
-                  query.find({
-                    success: function(results) {
-                      equal(results.length, 0);
-                      done();
-                    }
+                    var query = new Parse.Query(TestObject);
+                    query.find({
+                      success: function(results) {
+                        equal(results.length, 0);
+                        done();
+                      }
+                    });
                   });
-                });
               }
             });
           }
@@ -670,15 +670,15 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Update
-                  object.set("foo", "bar");
-                  object.save(null, {
-                    success: function() {
-                      done();
-                    }
+                    object.set("foo", "bar");
+                    object.save(null, {
+                      success: function() {
+                        done();
+                      }
+                    });
                   });
-                });
               }
             });
           }
@@ -716,14 +716,14 @@ describe('Parse.ACL', () => {
                 ok(object.get("ACL"));
 
                 Parse.User.logOut()
-                .then(() => {
+                  .then(() => {
                   // Delete
-                  object.destroy({
-                    success: function() {
-                      done();
-                    }
+                    object.destroy({
+                      success: function() {
+                        done();
+                      }
+                    });
                   });
-                });
               }
             });
           }
@@ -769,43 +769,43 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Sign in as Bob again.
-                  Parse.User.logIn("bob", "pass", {
-                    success: function() {
-                      var query = new Parse.Query(TestObject);
-                      query.get(object.id, {
-                        success: function(result) {
-                          ok(result);
-                          equal(result.id, object.id);
-                          done();
-                        }
-                      });
-                    }
-                  });
-                }
-              });
-            }
+                    // Sign in as Bob again.
+                    Parse.User.logIn("bob", "pass", {
+                      success: function() {
+                        var query = new Parse.Query(TestObject);
+                        query.get(object.id, {
+                          success: function(result) {
+                            ok(result);
+                            equal(result.id, object.id);
+                            done();
+                          }
+                        });
+                      }
+                    });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -815,49 +815,49 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Sign in as Bob again.
-                  Parse.User.logIn("bob", "pass", {
-                    success: function() {
-                      var query = new Parse.Query(TestObject);
-                      query.find({
-                        success: function(results) {
-                          equal(results.length, 1);
-                          var result = results[0];
-                          ok(result);
-                          if (!result) {
-                            fail("should have result");
-                          } else {
-                            equal(result.id, object.id);
+                    // Sign in as Bob again.
+                    Parse.User.logIn("bob", "pass", {
+                      success: function() {
+                        var query = new Parse.Query(TestObject);
+                        query.find({
+                          success: function(results) {
+                            equal(results.length, 1);
+                            var result = results[0];
+                            ok(result);
+                            if (!result) {
+                              fail("should have result");
+                            } else {
+                              equal(result.id, object.id);
+                            }
+                            done();
                           }
-                          done();
-                        }
-                      });
-                    }
-                  });
-                }
-              });
-            }
+                        });
+                      }
+                    });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -867,41 +867,41 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Sign in as Bob again.
-                  Parse.User.logIn("bob", "pass", {
-                    success: function() {
-                      object.set("foo", "bar");
-                      object.save(null, {
-                        success: function() {
-                          done();
-                        }
-                      });
-                    }
-                  });
-                }
-              });
-            }
+                    // Sign in as Bob again.
+                    Parse.User.logIn("bob", "pass", {
+                      success: function() {
+                        object.set("foo", "bar");
+                        object.save(null, {
+                          success: function() {
+                            done();
+                          }
+                        });
+                      }
+                    });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -911,41 +911,41 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Sign in as Bob again.
-                  Parse.User.logIn("bob", "pass", {
-                    success: function() {
-                      object.set("foo", "bar");
-                      object.destroy({
-                        success: function() {
-                          done();
-                        }
-                      });
-                    }
-                  });
-                }
-              });
-            }
+                    // Sign in as Bob again.
+                    Parse.User.logIn("bob", "pass", {
+                      success: function() {
+                        object.set("foo", "bar");
+                        object.destroy({
+                          success: function() {
+                            done();
+                          }
+                        });
+                      }
+                    });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -955,41 +955,41 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Start making requests by the public.
-                  Parse.User.logOut()
-                  .then(() => {
-                    var query = new Parse.Query(TestObject);
-                    query.get(object.id).then((result) => {
-                      fail(result);
-                    }, (error) => {
-                      expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
-                      done();
-                    });
-                  });
-                }
-              });
-            }
+                    // Start making requests by the public.
+                    Parse.User.logOut()
+                      .then(() => {
+                        var query = new Parse.Query(TestObject);
+                        query.get(object.id).then((result) => {
+                          fail(result);
+                        }, (error) => {
+                          expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
+                          done();
+                        });
+                      });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -999,41 +999,41 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Start making requests by the public.
-                  Parse.User.logOut()
-                  .then(() => {
-                    var query = new Parse.Query(TestObject);
-                    query.find({
-                      success: function(results) {
-                        equal(results.length, 0);
-                        done();
-                      }
-                    });
-                  });
-                }
-              });
-            }
+                    // Start making requests by the public.
+                    Parse.User.logOut()
+                      .then(() => {
+                        var query = new Parse.Query(TestObject);
+                        query.find({
+                          success: function(results) {
+                            equal(results.length, 0);
+                            done();
+                          }
+                        });
+                      });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -1043,41 +1043,41 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Start making requests by the public.
-                  Parse.User.logOut()
-                  .then(() => {
-                    object.set("foo", "bar");
-                    object.save().then(() => {
-                      fail('expected failure');
-                    }, (error) => {
-                      expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
-                      done();
-                    });
-                  });
-                }
-              });
-            }
+                    // Start making requests by the public.
+                    Parse.User.logOut()
+                      .then(() => {
+                        object.set("foo", "bar");
+                        object.save().then(() => {
+                          fail('expected failure');
+                        }, (error) => {
+                          expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
+                          done();
+                        });
+                      });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -1087,39 +1087,39 @@ describe('Parse.ACL', () => {
     Parse.User.signUp("bob", "pass", null, {
       success: function(bob) {
         Parse.User.logOut()
-        .then(() => {
+          .then(() => {
           // Sign in as Alice.
-          Parse.User.signUp("alice", "wonderland", null, {
-            success: function(alice) {
+            Parse.User.signUp("alice", "wonderland", null, {
+              success: function(alice) {
               // Create an object shared by Bob and Alice.
-              var object = new TestObject();
-              var acl = new Parse.ACL(alice);
-              acl.setWriteAccess(bob, true);
-              acl.setReadAccess(bob, true);
-              object.setACL(acl);
-              object.save(null, {
-                success: function() {
-                  equal(object.getACL().getReadAccess(alice), true);
-                  equal(object.getACL().getWriteAccess(alice), true);
-                  equal(object.getACL().getReadAccess(bob), true);
-                  equal(object.getACL().getWriteAccess(bob), true);
-                  equal(object.getACL().getPublicReadAccess(), false);
-                  equal(object.getACL().getPublicWriteAccess(), false);
+                var object = new TestObject();
+                var acl = new Parse.ACL(alice);
+                acl.setWriteAccess(bob, true);
+                acl.setReadAccess(bob, true);
+                object.setACL(acl);
+                object.save(null, {
+                  success: function() {
+                    equal(object.getACL().getReadAccess(alice), true);
+                    equal(object.getACL().getWriteAccess(alice), true);
+                    equal(object.getACL().getReadAccess(bob), true);
+                    equal(object.getACL().getWriteAccess(bob), true);
+                    equal(object.getACL().getPublicReadAccess(), false);
+                    equal(object.getACL().getPublicWriteAccess(), false);
 
-                  // Start making requests by the public.
-                  Parse.User.logOut()
-                  .then(() => object.destroy())
-                  .then(() => {
-                    fail('expected failure');
-                  }, (error) => {
-                    expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
-                    done();
-                  });
-                }
-              });
-            }
+                    // Start making requests by the public.
+                    Parse.User.logOut()
+                      .then(() => object.destroy())
+                      .then(() => {
+                        fail('expected failure');
+                      }, (error) => {
+                        expect(error.code).toEqual(Parse.Error.OBJECT_NOT_FOUND);
+                        done();
+                      });
+                  }
+                });
+              }
+            });
           });
-        });
       }
     });
   });
@@ -1172,18 +1172,18 @@ describe('Parse.ACL', () => {
     }, {
       success: function() {
         Parse.User.logOut()
-        .then(() => {
-          Parse.User.logIn("tdurden", "mayhem", {
-            success: function(user) {
-              equal(user.get("foo"), "bar");
-              done();
-            },
-            error: function(user, error) {
-              ok(null, "Error " + error.id + ": " + error.message);
-              done();
-            }
+          .then(() => {
+            Parse.User.logIn("tdurden", "mayhem", {
+              success: function(user) {
+                equal(user.get("foo"), "bar");
+                done();
+              },
+              error: function(user, error) {
+                ok(null, "Error " + error.id + ": " + error.message);
+                done();
+              }
+            });
           });
-        });
       },
       error: function(user, error) {
         ok(null, "Error " + error.id + ": " + error.message);
