@@ -33,14 +33,14 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var obj = results[0];
-      expect(obj.installationId).toEqual(installId);
-      expect(obj.deviceType).toEqual(device);
-      done();
-    }).catch((error) => { console.log(error); jfail(error); done(); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var obj = results[0];
+        expect(obj.installationId).toEqual(installId);
+        expect(obj.deviceType).toEqual(device);
+        done();
+      }).catch((error) => { console.log(error); jfail(error); done(); });
   });
 
   it('creates an ios installation with ids', (done) => {
@@ -51,14 +51,14 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var obj = results[0];
-      expect(obj.deviceToken).toEqual(t);
-      expect(obj.deviceType).toEqual(device);
-      done();
-    }).catch((error) => { console.log(error); jfail(error); done(); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var obj = results[0];
+        expect(obj.deviceToken).toEqual(t);
+        expect(obj.deviceType).toEqual(device);
+        done();
+      }).catch((error) => { console.log(error); jfail(error); done(); });
   });
 
   it('creates an embedded installation with ids', (done) => {
@@ -69,14 +69,14 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var obj = results[0];
-      expect(obj.installationId).toEqual(installId);
-      expect(obj.deviceType).toEqual(device);
-      done();
-    }).catch((error) => { console.log(error); jfail(error); done(); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var obj = results[0];
+        expect(obj.installationId).toEqual(installId);
+        expect(obj.deviceType).toEqual(device);
+        done();
+      }).catch((error) => { console.log(error); jfail(error); done(); });
   });
 
   it('creates an android installation with all fields', (done) => {
@@ -88,18 +88,18 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var obj = results[0];
-      expect(obj.installationId).toEqual(installId);
-      expect(obj.deviceType).toEqual(device);
-      expect(typeof obj.channels).toEqual('object');
-      expect(obj.channels.length).toEqual(2);
-      expect(obj.channels[0]).toEqual('foo');
-      expect(obj.channels[1]).toEqual('bar');
-      done();
-    }).catch((error) => { console.log(error); jfail(error); done(); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var obj = results[0];
+        expect(obj.installationId).toEqual(installId);
+        expect(obj.deviceType).toEqual(device);
+        expect(typeof obj.channels).toEqual('object');
+        expect(obj.channels.length).toEqual(2);
+        expect(obj.channels[0]).toEqual('foo');
+        expect(obj.channels[1]).toEqual('bar');
+        done();
+      }).catch((error) => { console.log(error); jfail(error); done(); });
   });
 
   it('creates an ios installation with all fields', (done) => {
@@ -111,18 +111,18 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var obj = results[0];
-      expect(obj.deviceToken).toEqual(t);
-      expect(obj.deviceType).toEqual(device);
-      expect(typeof obj.channels).toEqual('object');
-      expect(obj.channels.length).toEqual(2);
-      expect(obj.channels[0]).toEqual('foo');
-      expect(obj.channels[1]).toEqual('bar');
-      done();
-    }).catch((error) => { console.log(error); jfail(error); done(); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var obj = results[0];
+        expect(obj.deviceToken).toEqual(t);
+        expect(obj.deviceType).toEqual(device);
+        expect(typeof obj.channels).toEqual('object');
+        expect(obj.channels.length).toEqual(2);
+        expect(obj.channels[0]).toEqual('foo');
+        expect(obj.channels[1]).toEqual('bar');
+        done();
+      }).catch((error) => { console.log(error); jfail(error); done(); });
   });
 
   it('should properly fail queying installations', (done) => {
@@ -133,17 +133,17 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      const query = new Parse.Query(Parse.Installation);
-      return query.find()
-    }).then(() => {
-      fail('Should not succeed!');
-      done();
-    }).catch((error) => {
-      expect(error.code).toBe(119);
-      expect(error.message).toBe('Clients aren\'t allowed to perform the find operation on the installation collection.')
-      done();
-    });
+      .then(() => {
+        const query = new Parse.Query(Parse.Installation);
+        return query.find()
+      }).then(() => {
+        fail('Should not succeed!');
+        done();
+      }).catch((error) => {
+        expect(error.code).toBe(119);
+        expect(error.message).toBe('Clients aren\'t allowed to perform the find operation on the installation collection.')
+        done();
+      });
   });
 
   it('should properly queying installations with masterKey', (done) => {
@@ -154,19 +154,19 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      const query = new Parse.Query(Parse.Installation);
-      return query.find({useMasterKey: true});
-    }).then((results) => {
-      expect(results.length).toEqual(1);
-      var obj = results[0].toJSON();
-      expect(obj.installationId).toEqual(installId);
-      expect(obj.deviceType).toEqual(device);
-      done();
-    }).catch(() => {
-      fail('Should not fail');
-      done();
-    });
+      .then(() => {
+        const query = new Parse.Query(Parse.Installation);
+        return query.find({useMasterKey: true});
+      }).then((results) => {
+        expect(results.length).toEqual(1);
+        var obj = results[0].toJSON();
+        expect(obj.installationId).toEqual(installId);
+        expect(obj.deviceType).toEqual(device);
+        done();
+      }).catch(() => {
+        fail('Should not fail');
+        done();
+      });
   });
 
   it('fails with missing ids', (done) => {
@@ -175,13 +175,13 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      fail('Should not have been able to create an Installation.');
-      done();
-    }).catch((error) => {
-      expect(error.code).toEqual(135);
-      done();
-    });
+      .then(() => {
+        fail('Should not have been able to create an Installation.');
+        done();
+      }).catch((error) => {
+        expect(error.code).toEqual(135);
+        done();
+      });
   });
 
   it('fails for android with missing type', (done) => {
@@ -191,13 +191,13 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      fail('Should not have been able to create an Installation.');
-      done();
-    }).catch((error) => {
-      expect(error.code).toEqual(135);
-      done();
-    });
+      .then(() => {
+        fail('Should not have been able to create an Installation.');
+        done();
+      }).catch((error) => {
+        expect(error.code).toEqual(135);
+        done();
+      });
   });
 
   it('creates an object with custom fields', (done) => {
@@ -209,13 +209,13 @@ describe('Installations', () => {
       'custom': 'allowed'
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var obj = results[0];
-      expect(obj.custom).toEqual('allowed');
-      done();
-    }).catch((error) => { console.log(error); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var obj = results[0];
+        expect(obj.custom).toEqual('allowed');
+        done();
+      }).catch((error) => { console.log(error); });
   });
 
   // Note: did not port test 'TestObjectIDForIdentifiers'
@@ -232,24 +232,24 @@ describe('Installations', () => {
     var firstObject;
     var secondObject;
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      firstObject = results[0];
-      delete input.deviceToken;
-      delete input.channels;
-      input['foo'] = 'bar';
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      secondObject = results[0];
-      expect(firstObject._id).toEqual(secondObject._id);
-      expect(secondObject.channels.length).toEqual(2);
-      expect(secondObject.foo).toEqual('bar');
-      done();
-    }).catch((error) => { console.log(error); });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        firstObject = results[0];
+        delete input.deviceToken;
+        delete input.channels;
+        input['foo'] = 'bar';
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        secondObject = results[0];
+        expect(firstObject._id).toEqual(secondObject._id);
+        expect(secondObject.channels.length).toEqual(2);
+        expect(secondObject.foo).toEqual('bar');
+        done();
+      }).catch((error) => { console.log(error); });
   });
 
   it('merging when two objects both only have one id', (done) => {
@@ -271,31 +271,31 @@ describe('Installations', () => {
     var firstObject;
     var secondObject;
     rest.create(config, auth.nobody(config), '_Installation', input1)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      firstObject = results[0];
-      return rest.create(config, auth.nobody(config), '_Installation', input2);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(2);
-      if (results[0]['_id'] == firstObject._id) {
-        secondObject = results[1];
-      } else {
-        secondObject = results[0];
-      }
-      return rest.create(config, auth.nobody(config), '_Installation', input3);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0]['_id']).toEqual(secondObject._id);
-      done();
-    }).catch((error) => {
-      jfail(error);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        firstObject = results[0];
+        return rest.create(config, auth.nobody(config), '_Installation', input2);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(2);
+        if (results[0]['_id'] == firstObject._id) {
+          secondObject = results[1];
+        } else {
+          secondObject = results[0];
+        }
+        return rest.create(config, auth.nobody(config), '_Installation', input3);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0]['_id']).toEqual(secondObject._id);
+        done();
+      }).catch((error) => {
+        jfail(error);
+        done();
+      });
   });
 
   xit('creating multiple devices with same device token works', (done) => {
@@ -309,24 +309,24 @@ describe('Installations', () => {
       'deviceToken': t
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      input.installationId = installId2;
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    }).then(() => {
-      input.installationId = installId3;
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', {installationId: installId1}, installationSchema, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      return database.adapter.find('_Installation', {installationId: installId2}, installationSchema, {});
-    }).then(results => {
-      expect(results.length).toEqual(1);
-      return database.adapter.find('_Installation', {installationId: installId3}, installationSchema, {});
-    }).then((results) => {
-      expect(results.length).toEqual(1);
-      done();
-    }).catch((error) => { console.log(error); });
+      .then(() => {
+        input.installationId = installId2;
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      }).then(() => {
+        input.installationId = installId3;
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', {installationId: installId1}, installationSchema, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        return database.adapter.find('_Installation', {installationId: installId2}, installationSchema, {});
+      }).then(results => {
+        expect(results.length).toEqual(1);
+        return database.adapter.find('_Installation', {installationId: installId3}, installationSchema, {});
+      }).then((results) => {
+        expect(results.length).toEqual(1);
+        done();
+      }).catch((error) => { console.log(error); });
   });
 
   it('updating with new channels', (done) => {
@@ -336,25 +336,25 @@ describe('Installations', () => {
       channels: ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      var objectId = results[0].objectId;
-      var update = {
-        'channels': ['baz']
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId }, update);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].channels.length).toEqual(1);
-      expect(results[0].channels[0]).toEqual('baz');
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        var objectId = results[0].objectId;
+        var update = {
+          'channels': ['baz']
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId }, update);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].channels.length).toEqual(1);
+        expect(results[0].channels[0]).toEqual('baz');
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update android fails with new installation id', (done) => {
@@ -366,18 +366,18 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = { 'installationId': installId2 };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    }).then(() => {
-      fail('Updating the installation should have failed.');
-      done();
-    }).catch((error) => {
-      expect(error.code).toEqual(136);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = { 'installationId': installId2 };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      }).then(() => {
+        fail('Updating the installation should have failed.');
+        done();
+      }).catch((error) => {
+        expect(error.code).toEqual(136);
+        done();
+      });
   });
 
   it('update ios fails with new deviceToken and no installationId', (done) => {
@@ -389,17 +389,17 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = { 'deviceToken': b };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    }).then(() => {
-      fail('Updating the installation should have failed.');
-    }).catch((error) => {
-      expect(error.code).toEqual(136);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = { 'deviceToken': b };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      }).then(() => {
+        fail('Updating the installation should have failed.');
+      }).catch((error) => {
+        expect(error.code).toEqual(136);
+        done();
+      });
   });
 
   it('update ios updates device token', (done) => {
@@ -413,25 +413,25 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'installationId': installId,
-        'deviceToken': u,
-        'deviceType': 'ios'
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].deviceToken).toEqual(u);
-      done();
-    }).catch(err => {
-      jfail(err);
-      done();
-    })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'installationId': installId,
+          'deviceToken': u,
+          'deviceType': 'ios'
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].deviceToken).toEqual(u);
+        done();
+      }).catch(err => {
+        jfail(err);
+        done();
+      })
   });
 
   it('update fails to change deviceType', (done) => {
@@ -442,20 +442,20 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'deviceType': 'ios'
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    }).then(() => {
-      fail('Should not have been able to update Installation.');
-      done();
-    }).catch((error) => {
-      expect(error.code).toEqual(136);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'deviceType': 'ios'
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      }).then(() => {
+        fail('Should not have been able to update Installation.');
+        done();
+      }).catch((error) => {
+        expect(error.code).toEqual(136);
+        done();
+      });
   });
 
   it('update android with custom field', (done) => {
@@ -466,20 +466,20 @@ describe('Installations', () => {
       'channels': ['foo', 'bar']
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'custom': 'allowed'
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0]['custom']).toEqual('allowed');
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'custom': 'allowed'
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0]['custom']).toEqual('allowed');
+        done();
+      });
   });
 
   it('update android device token with duplicate device token', (done) => {
@@ -494,37 +494,37 @@ describe('Installations', () => {
     var firstObject;
     var secondObject;
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      input = {
-        'installationId': installId2,
-        'deviceType': 'android'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {installationId: installId1}, {}))
-    .then(results => {
-      firstObject = results[0];
-      expect(results.length).toEqual(1);
-      return database.adapter.find('_Installation', installationSchema, {installationId: installId2}, {});
-    }).then(results => {
-      expect(results.length).toEqual(1);
-      secondObject = results[0];
-      // Update second installation to conflict with first installation
-      input = {
-        'objectId': secondObject.objectId,
-        'deviceToken': t
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: secondObject.objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {objectId: firstObject.objectId}, {}))
-    .then(results => {
+      .then(() => {
+        input = {
+          'installationId': installId2,
+          'deviceType': 'android'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {installationId: installId1}, {}))
+      .then(results => {
+        firstObject = results[0];
+        expect(results.length).toEqual(1);
+        return database.adapter.find('_Installation', installationSchema, {installationId: installId2}, {});
+      }).then(results => {
+        expect(results.length).toEqual(1);
+        secondObject = results[0];
+        // Update second installation to conflict with first installation
+        input = {
+          'objectId': secondObject.objectId,
+          'deviceToken': t
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: secondObject.objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {objectId: firstObject.objectId}, {}))
+      .then(results => {
       // The first object should have been deleted
-      expect(results.length).toEqual(0);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+        expect(results.length).toEqual(0);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update ios device token with duplicate device token', (done) => {
@@ -539,41 +539,41 @@ describe('Installations', () => {
     var firstObject;
     var secondObject;
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      input = {
-        'installationId': installId2,
-        'deviceType': 'ios'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => delay(100))
-    .then(() => database.adapter.find('_Installation', installationSchema, {installationId: installId1}, {}))
-    .then((results) => {
-      expect(results.length).toEqual(1);
-      firstObject = results[0];
-    })
-    .then(() => delay(100))
-    .then(() => database.adapter.find('_Installation', installationSchema, {installationId: installId2}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      secondObject = results[0];
-      // Update second installation to conflict with first installation id
-      input = {
-        'installationId': installId2,
-        'deviceToken': t
-      };
-      return rest.update(config, auth.nobody(config), '_Installation',  { objectId: secondObject.objectId }, input);
-    })
-    .then(() => delay(100))
-    .then(() => database.adapter.find('_Installation', installationSchema, {objectId: firstObject.objectId}, {}))
-    .then(results => {
+      .then(() => {
+        input = {
+          'installationId': installId2,
+          'deviceType': 'ios'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => delay(100))
+      .then(() => database.adapter.find('_Installation', installationSchema, {installationId: installId1}, {}))
+      .then((results) => {
+        expect(results.length).toEqual(1);
+        firstObject = results[0];
+      })
+      .then(() => delay(100))
+      .then(() => database.adapter.find('_Installation', installationSchema, {installationId: installId2}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        secondObject = results[0];
+        // Update second installation to conflict with first installation id
+        input = {
+          'installationId': installId2,
+          'deviceToken': t
+        };
+        return rest.update(config, auth.nobody(config), '_Installation',  { objectId: secondObject.objectId }, input);
+      })
+      .then(() => delay(100))
+      .then(() => database.adapter.find('_Installation', installationSchema, {objectId: firstObject.objectId}, {}))
+      .then(results => {
       // The first object should have been deleted
-      expect(results.length).toEqual(0);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+        expect(results.length).toEqual(0);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   xit('update ios device token with duplicate token different app', (done) => {
@@ -587,21 +587,21 @@ describe('Installations', () => {
       'appIdentifier': 'foo'
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      input.installationId = installId2;
-      input.appIdentifier = 'bar';
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
+      .then(() => {
+        input.installationId = installId2;
+        input.appIdentifier = 'bar';
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
       // The first object should have been deleted during merge
-      expect(results.length).toEqual(1);
-      expect(results[0].installationId).toEqual(installId2);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+        expect(results.length).toEqual(1);
+        expect(results[0].installationId).toEqual(installId2);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update ios token and channels', (done) => {
@@ -612,26 +612,26 @@ describe('Installations', () => {
       'deviceType': 'ios'
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'deviceToken': t,
-        'channels': []
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].installationId).toEqual(installId);
-      expect(results[0].deviceToken).toEqual(t);
-      expect(results[0].channels.length).toEqual(0);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'deviceToken': t,
+          'channels': []
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].installationId).toEqual(installId);
+        expect(results[0].deviceToken).toEqual(t);
+        expect(results[0].channels.length).toEqual(0);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update ios linking two existing objects', (done) => {
@@ -642,34 +642,34 @@ describe('Installations', () => {
       'deviceType': 'ios'
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      input = {
-        'deviceToken': t,
-        'deviceType': 'ios'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'deviceToken': t,
-        'installationId': installId,
-        'deviceType': 'ios'
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].installationId).toEqual(installId);
-      expect(results[0].deviceToken).toEqual(t);
-      expect(results[0].deviceType).toEqual('ios');
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => {
+        input = {
+          'deviceToken': t,
+          'deviceType': 'ios'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'deviceToken': t,
+          'installationId': installId,
+          'deviceType': 'ios'
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].installationId).toEqual(installId);
+        expect(results[0].deviceToken).toEqual(t);
+        expect(results[0].deviceType).toEqual('ios');
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update is linking two existing objects w/ increment', (done) => {
@@ -680,39 +680,39 @@ describe('Installations', () => {
       'deviceType': 'ios'
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      input = {
-        'deviceToken': t,
-        'deviceType': 'ios'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'deviceToken': t,
-        'installationId': installId,
-        'deviceType': 'ios',
-        'score': {
-          '__op': 'Increment',
-          'amount': 1
-        }
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].installationId).toEqual(installId);
-      expect(results[0].deviceToken).toEqual(t);
-      expect(results[0].deviceType).toEqual('ios');
-      expect(results[0].score).toEqual(1);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => {
+        input = {
+          'deviceToken': t,
+          'deviceType': 'ios'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'deviceToken': t,
+          'installationId': installId,
+          'deviceType': 'ios',
+          'score': {
+            '__op': 'Increment',
+            'amount': 1
+          }
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: results[0].objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].installationId).toEqual(installId);
+        expect(results[0].deviceToken).toEqual(t);
+        expect(results[0].deviceType).toEqual('ios');
+        expect(results[0].score).toEqual(1);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update is linking two existing with installation id', (done) => {
@@ -725,37 +725,37 @@ describe('Installations', () => {
     var installObj;
     var tokenObj;
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      installObj = results[0];
-      input = {
-        'deviceToken': t,
-        'deviceType': 'ios'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      tokenObj = results[0];
-      input = {
-        'installationId': installId,
-        'deviceToken': t,
-        'deviceType': 'ios'
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: installObj.objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, { objectId: tokenObj.objectId }, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].installationId).toEqual(installId);
-      expect(results[0].deviceToken).toEqual(t);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        installObj = results[0];
+        input = {
+          'deviceToken': t,
+          'deviceType': 'ios'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        tokenObj = results[0];
+        input = {
+          'installationId': installId,
+          'deviceToken': t,
+          'deviceType': 'ios'
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: installObj.objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, { objectId: tokenObj.objectId }, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].installationId).toEqual(installId);
+        expect(results[0].deviceToken).toEqual(t);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('update is linking two existing with installation id w/ op', (done) => {
@@ -768,42 +768,42 @@ describe('Installations', () => {
     var installObj;
     var tokenObj;
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      installObj = results[0];
-      input = {
-        'deviceToken': t,
-        'deviceType': 'ios'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      tokenObj = results[0];
-      input = {
-        'installationId': installId,
-        'deviceToken': t,
-        'deviceType': 'ios',
-        'score': {
-          '__op': 'Increment',
-          'amount': 1
-        }
-      };
-      return rest.update(config, auth.nobody(config), '_Installation', { objectId: installObj.objectId }, input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, { objectId: tokenObj.objectId }, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].installationId).toEqual(installId);
-      expect(results[0].deviceToken).toEqual(t);
-      expect(results[0].score).toEqual(1);
-      done();
-    }).catch(error => {
-      jfail(error);
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        installObj = results[0];
+        input = {
+          'deviceToken': t,
+          'deviceType': 'ios'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, { deviceToken: t }, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        tokenObj = results[0];
+        input = {
+          'installationId': installId,
+          'deviceToken': t,
+          'deviceType': 'ios',
+          'score': {
+            '__op': 'Increment',
+            'amount': 1
+          }
+        };
+        return rest.update(config, auth.nobody(config), '_Installation', { objectId: installObj.objectId }, input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, { objectId: tokenObj.objectId }, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].installationId).toEqual(installId);
+        expect(results[0].deviceToken).toEqual(t);
+        expect(results[0].score).toEqual(1);
+        done();
+      }).catch(error => {
+        jfail(error);
+        done();
+      });
   });
 
   it('ios merge existing same token no installation id', (done) => {
@@ -824,28 +824,28 @@ describe('Installations', () => {
       'deviceType': 'ios'
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      input = {
-        'installationId': installId,
-        'deviceToken': t,
-        'deviceType': 'ios'
-      };
-      return rest.create(config, auth.nobody(config), '_Installation', input);
-    })
-    .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
-    .then(results => {
-      expect(results.length).toEqual(1);
-      expect(results[0].deviceToken).toEqual(t);
-      expect(results[0].installationId).toEqual(installId);
-      done();
-    })
-    .catch(error => {
-      console.log(error);
-      fail();
-      done();
-    });
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        input = {
+          'installationId': installId,
+          'deviceToken': t,
+          'deviceType': 'ios'
+        };
+        return rest.create(config, auth.nobody(config), '_Installation', input);
+      })
+      .then(() => database.adapter.find('_Installation', installationSchema, {}, {}))
+      .then(results => {
+        expect(results.length).toEqual(1);
+        expect(results[0].deviceToken).toEqual(t);
+        expect(results[0].installationId).toEqual(installId);
+        done();
+      })
+      .catch(error => {
+        console.log(error);
+        fail();
+        done();
+      });
   });
 
   it('allows you to get your own installation (regression test for #1718)', done => {
@@ -856,25 +856,25 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(createResult => {
-      const headers = {
-        'X-Parse-Application-Id': 'test',
-        'X-Parse-REST-API-Key':   'rest',
-      };
-      request.get({
-        headers: headers,
-        url: 'http://localhost:8378/1/installations/' + createResult.response.objectId,
-        json: true,
-      }, (error, response, body) => {
-        expect(body.objectId).toEqual(createResult.response.objectId);
+      .then(createResult => {
+        const headers = {
+          'X-Parse-Application-Id': 'test',
+          'X-Parse-REST-API-Key':   'rest',
+        };
+        request.get({
+          headers: headers,
+          url: 'http://localhost:8378/1/installations/' + createResult.response.objectId,
+          json: true,
+        }, (error, response, body) => {
+          expect(body.objectId).toEqual(createResult.response.objectId);
+          done();
+        });
+      })
+      .catch(error => {
+        console.log(error);
+        fail('failed');
         done();
       });
-    })
-    .catch(error => {
-      console.log(error);
-      fail('failed');
-      done();
-    });
   });
 
   it('allows you to update installation from header (#2090)', done => {
@@ -885,30 +885,30 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(() => {
-      const headers = {
-        'X-Parse-Application-Id': 'test',
-        'X-Parse-REST-API-Key':   'rest',
-        'X-Parse-Installation-Id': installId
-      };
-      request.post({
-        headers: headers,
-        url: 'http://localhost:8378/1/classes/_Installation',
-        json: true,
-        body: {
-          date: new Date()
-        }
-      }, (error, response, body) => {
-        expect(response.statusCode).toBe(200);
-        expect(body.updatedAt).not.toBeUndefined();
+      .then(() => {
+        const headers = {
+          'X-Parse-Application-Id': 'test',
+          'X-Parse-REST-API-Key':   'rest',
+          'X-Parse-Installation-Id': installId
+        };
+        request.post({
+          headers: headers,
+          url: 'http://localhost:8378/1/classes/_Installation',
+          json: true,
+          body: {
+            date: new Date()
+          }
+        }, (error, response, body) => {
+          expect(response.statusCode).toBe(200);
+          expect(body.updatedAt).not.toBeUndefined();
+          done();
+        });
+      })
+      .catch(error => {
+        console.log(error);
+        fail('failed');
         done();
       });
-    })
-    .catch(error => {
-      console.log(error);
-      fail('failed');
-      done();
-    });
   });
 
   it('allows you to update installation with masterKey', done => {
@@ -919,19 +919,19 @@ describe('Installations', () => {
       'deviceType': device
     };
     rest.create(config, auth.nobody(config), '_Installation', input)
-    .then(createResult => {
-      const installationObj = Parse.Installation.createWithoutData(createResult.response.objectId);
-      installationObj.set('customField', 'custom value');
-      return installationObj.save(null, {useMasterKey: true});
-    }).then(updateResult => {
-      expect(updateResult).not.toBeUndefined();
-      expect(updateResult.get('customField')).toEqual('custom value');
-      done();
-    }).catch(error => {
-      console.log(error);
-      fail('failed');
-      done();
-    });
+      .then(createResult => {
+        const installationObj = Parse.Installation.createWithoutData(createResult.response.objectId);
+        installationObj.set('customField', 'custom value');
+        return installationObj.save(null, {useMasterKey: true});
+      }).then(updateResult => {
+        expect(updateResult).not.toBeUndefined();
+        expect(updateResult.get('customField')).toEqual('custom value');
+        done();
+      }).catch(error => {
+        console.log(error);
+        fail('failed');
+        done();
+      });
   });
 
   it('should properly handle installation save #2780', done => {
