@@ -61,10 +61,10 @@ describe('SchemaController', () => {
       return schema.validateObject('Stuff', {aRelation: {__type:'Relation',className:'Stuff'}});
     }).then((schema) => {
       return schema.validateObject('Stuff', {aRelation: {__type:'Pointer',className:'Stuff'}})
-      .then(() => {
-        fail('expected invalidity');
-        done();
-      }, done);
+        .then(() => {
+          fail('expected invalidity');
+          done();
+        }, done);
     }, (err) => {
       fail(err);
       done();
@@ -142,40 +142,40 @@ describe('SchemaController', () => {
   it('class-level permissions test get', (done) => {
     var obj;
     createTestUser()
-    .then(user => {
-      return config.database.loadSchema()
-      // Create a valid class
-      .then(schema => schema.validateObject('Stuff', {foo: 'bar'}))
-      .then(schema => {
-        var find = {};
-        var get = {};
-        get[user.id] = true;
-        return schema.setPermissions('Stuff', {
-          'create': {'*': true},
-          'find': find,
-          'get': get
-        });
-      }).then(() => {
-        obj = new Parse.Object('Stuff');
-        obj.set('foo', 'bar');
-        return obj.save();
-      }).then((o) => {
-        obj = o;
-        var query = new Parse.Query('Stuff');
-        return query.find();
-      }).then(() => {
-        fail('Class permissions should have rejected this query.');
-        done();
-      }, () => {
-        var query = new Parse.Query('Stuff');
-        return query.get(obj.id).then(() => {
-          done();
-        }, () => {
-          fail('Class permissions should have allowed this get query');
-          done();
-        });
+      .then(user => {
+        return config.database.loadSchema()
+        // Create a valid class
+          .then(schema => schema.validateObject('Stuff', {foo: 'bar'}))
+          .then(schema => {
+            var find = {};
+            var get = {};
+            get[user.id] = true;
+            return schema.setPermissions('Stuff', {
+              'create': {'*': true},
+              'find': find,
+              'get': get
+            });
+          }).then(() => {
+            obj = new Parse.Object('Stuff');
+            obj.set('foo', 'bar');
+            return obj.save();
+          }).then((o) => {
+            obj = o;
+            var query = new Parse.Query('Stuff');
+            return query.find();
+          }).then(() => {
+            fail('Class permissions should have rejected this query.');
+            done();
+          }, () => {
+            var query = new Parse.Query('Stuff');
+            return query.get(obj.id).then(() => {
+              done();
+            }, () => {
+              fail('Class permissions should have allowed this get query');
+              done();
+            });
+          });
       });
-    });
   });
 
   it('class-level permissions test count', (done) => {
@@ -212,34 +212,34 @@ describe('SchemaController', () => {
 
   it('can add classes without needing an object', done => {
     config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'String'},
-    }))
-    .then(actualSchema => {
-      const expectedSchema = {
-        className: 'NewClass',
-        fields: {
-          objectId: { type: 'String' },
-          updatedAt: { type: 'Date' },
-          createdAt: { type: 'Date' },
-          ACL: { type: 'ACL' },
-          foo: { type: 'String' },
-        },
-        classLevelPermissions: {
-          find: { '*': true },
-          get: { '*': true },
-          create: { '*': true },
-          update: { '*': true },
-          delete: { '*': true },
-          addField: { '*': true },
-        },
-      }
-      expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      done();
-    })
-    .catch(error => {
-      fail('Error creating class: ' + JSON.stringify(error));
-    });
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'String'},
+      }))
+      .then(actualSchema => {
+        const expectedSchema = {
+          className: 'NewClass',
+          fields: {
+            objectId: { type: 'String' },
+            updatedAt: { type: 'Date' },
+            createdAt: { type: 'Date' },
+            ACL: { type: 'ACL' },
+            foo: { type: 'String' },
+          },
+          classLevelPermissions: {
+            find: { '*': true },
+            get: { '*': true },
+            create: { '*': true },
+            update: { '*': true },
+            delete: { '*': true },
+            addField: { '*': true },
+          },
+        }
+        expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
+        done();
+      })
+      .catch(error => {
+        fail('Error creating class: ' + JSON.stringify(error));
+      });
   });
 
   it('can update classes without needing an object', done => {
@@ -337,10 +337,195 @@ describe('SchemaController', () => {
     // If two callers race to create the same schema, the response to the
     // race loser should be the same as if they hadn't been racing.
     config.database.loadSchema()
-    .then(schema => {
-      var p1 = schema.addClassIfNotExists('NewClass', {foo: {type: 'String'}});
-      var p2 = schema.addClassIfNotExists('NewClass', {foo: {type: 'String'}});
-      Promise.race([p1, p2])
+      .then(schema => {
+        var p1 = schema.addClassIfNotExists('NewClass', {foo: {type: 'String'}});
+        var p2 = schema.addClassIfNotExists('NewClass', {foo: {type: 'String'}});
+        Promise.race([p1, p2])
+          .then(actualSchema => {
+            const expectedSchema = {
+              className: 'NewClass',
+              fields: {
+                objectId: { type: 'String' },
+                updatedAt: { type: 'Date' },
+                createdAt: { type: 'Date' },
+                ACL: { type: 'ACL' },
+                foo: { type: 'String' },
+              },
+              classLevelPermissions: {
+                find: { '*': true },
+                get: { '*': true },
+                create: { '*': true },
+                update: { '*': true },
+                delete: { '*': true },
+                addField: { '*': true },
+              },
+            }
+            expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
+          });
+        Promise.all([p1,p2])
+          .catch(error => {
+            expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
+            expect(error.message).toEqual('Class NewClass already exists.');
+            done();
+          });
+      });
+  });
+
+  it('refuses to create classes with invalid names', done => {
+    config.database.loadSchema()
+      .then(schema => {
+        schema.addClassIfNotExists('_InvalidName', {foo: {type: 'String'}})
+          .catch(error => {
+            expect(error.error).toEqual(
+              'Invalid classname: _InvalidName, classnames can only have alphanumeric characters and _, and must start with an alpha character '
+            );
+            done();
+          });
+      });
+  });
+
+  it('refuses to add fields with invalid names', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {'0InvalidName': {type: 'String'}}))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_KEY_NAME);
+        expect(error.error).toEqual('invalid field name: 0InvalidName');
+        done();
+      });
+  });
+
+  it('refuses to explicitly create the default fields for custom classes', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {objectId: {type: 'String'}}))
+      .catch(error => {
+        expect(error.code).toEqual(136);
+        expect(error.error).toEqual('field objectId cannot be added');
+        done();
+      });
+  });
+
+  it('refuses to explicitly create the default fields for non-custom classes', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('_Installation', {localeIdentifier: {type: 'String'}}))
+      .catch(error => {
+        expect(error.code).toEqual(136);
+        expect(error.error).toEqual('field localeIdentifier cannot be added');
+        done();
+      });
+  });
+
+  it('refuses to add fields with invalid types', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 7}
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_JSON);
+        expect(error.error).toEqual('invalid JSON');
+        done();
+      });
+  });
+
+  it('refuses to add fields with invalid pointer types', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Pointer'}
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(135);
+        expect(error.error).toEqual('type Pointer needs a class name');
+        done();
+      });
+  });
+
+  it('refuses to add fields with invalid pointer target', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Pointer', targetClass: 7},
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_JSON);
+        expect(error.error).toEqual('invalid JSON');
+        done();
+      });
+  });
+
+  it('refuses to add fields with invalid Relation type', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Relation', uselessKey: 7},
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(135);
+        expect(error.error).toEqual('type Relation needs a class name');
+        done();
+      });
+  });
+
+  it('refuses to add fields with invalid relation target', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Relation', targetClass: 7},
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_JSON);
+        expect(error.error).toEqual('invalid JSON');
+        done();
+      });
+  });
+
+  it('refuses to add fields with uncreatable pointer target class', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Pointer', targetClass: 'not a valid class name'},
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
+        expect(error.error).toEqual('Invalid classname: not a valid class name, classnames can only have alphanumeric characters and _, and must start with an alpha character ');
+        done();
+      });
+  });
+
+  it('refuses to add fields with uncreatable relation target class', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Relation', targetClass: 'not a valid class name'},
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
+        expect(error.error).toEqual('Invalid classname: not a valid class name, classnames can only have alphanumeric characters and _, and must start with an alpha character ');
+        done();
+      });
+  });
+
+  it('refuses to add fields with unknown types', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        foo: {type: 'Unknown'},
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INCORRECT_TYPE);
+        expect(error.error).toEqual('invalid field type: Unknown');
+        done();
+      });
+  });
+
+  it('will create classes', done => {
+    config.database.loadSchema()
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        aNumber: {type: 'Number'},
+        aString: {type: 'String'},
+        aBool: {type: 'Boolean'},
+        aDate: {type: 'Date'},
+        aObject: {type: 'Object'},
+        aArray: {type: 'Array'},
+        aGeoPoint: {type: 'GeoPoint'},
+        aFile: {type: 'File'},
+        aPointer: {type: 'Pointer', targetClass: 'ThisClassDoesNotExistYet'},
+        aRelation: {type: 'Relation', targetClass: 'NewClass'},
+        aBytes: {type: 'Bytes'},
+        aPolygon: {type: 'Polygon'},
+      }))
       .then(actualSchema => {
         const expectedSchema = {
           className: 'NewClass',
@@ -349,7 +534,18 @@ describe('SchemaController', () => {
             updatedAt: { type: 'Date' },
             createdAt: { type: 'Date' },
             ACL: { type: 'ACL' },
-            foo: { type: 'String' },
+            aString: { type: 'String' },
+            aNumber: { type: 'Number' },
+            aBool: { type: 'Boolean' },
+            aDate: { type: 'Date' },
+            aObject: { type: 'Object' },
+            aArray: { type: 'Array' },
+            aGeoPoint: { type: 'GeoPoint' },
+            aFile: { type: 'File' },
+            aPointer: { type: 'Pointer', targetClass: 'ThisClassDoesNotExistYet' },
+            aRelation: { type: 'Relation', targetClass: 'NewClass' },
+            aBytes: {type: 'Bytes'},
+            aPolygon: {type: 'Polygon'},
           },
           classLevelPermissions: {
             find: { '*': true },
@@ -361,395 +557,201 @@ describe('SchemaController', () => {
           },
         }
         expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      });
-      Promise.all([p1,p2])
-      .catch(error => {
-        expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
-        expect(error.message).toEqual('Class NewClass already exists.');
         done();
       });
-    });
-  });
-
-  it('refuses to create classes with invalid names', done => {
-    config.database.loadSchema()
-    .then(schema => {
-      schema.addClassIfNotExists('_InvalidName', {foo: {type: 'String'}})
-      .catch(error => {
-        expect(error.error).toEqual(
-          'Invalid classname: _InvalidName, classnames can only have alphanumeric characters and _, and must start with an alpha character '
-        );
-        done();
-      });
-    });
-  });
-
-  it('refuses to add fields with invalid names', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {'0InvalidName': {type: 'String'}}))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_KEY_NAME);
-      expect(error.error).toEqual('invalid field name: 0InvalidName');
-      done();
-    });
-  });
-
-  it('refuses to explicitly create the default fields for custom classes', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {objectId: {type: 'String'}}))
-    .catch(error => {
-      expect(error.code).toEqual(136);
-      expect(error.error).toEqual('field objectId cannot be added');
-      done();
-    });
-  });
-
-  it('refuses to explicitly create the default fields for non-custom classes', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('_Installation', {localeIdentifier: {type: 'String'}}))
-    .catch(error => {
-      expect(error.code).toEqual(136);
-      expect(error.error).toEqual('field localeIdentifier cannot be added');
-      done();
-    });
-  });
-
-  it('refuses to add fields with invalid types', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 7}
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_JSON);
-      expect(error.error).toEqual('invalid JSON');
-      done();
-    });
-  });
-
-  it('refuses to add fields with invalid pointer types', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Pointer'}
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(135);
-      expect(error.error).toEqual('type Pointer needs a class name');
-      done();
-    });
-  });
-
-  it('refuses to add fields with invalid pointer target', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Pointer', targetClass: 7},
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_JSON);
-      expect(error.error).toEqual('invalid JSON');
-      done();
-    });
-  });
-
-  it('refuses to add fields with invalid Relation type', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Relation', uselessKey: 7},
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(135);
-      expect(error.error).toEqual('type Relation needs a class name');
-      done();
-    });
-  });
-
-  it('refuses to add fields with invalid relation target', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Relation', targetClass: 7},
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_JSON);
-      expect(error.error).toEqual('invalid JSON');
-      done();
-    });
-  });
-
-  it('refuses to add fields with uncreatable pointer target class', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Pointer', targetClass: 'not a valid class name'},
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
-      expect(error.error).toEqual('Invalid classname: not a valid class name, classnames can only have alphanumeric characters and _, and must start with an alpha character ');
-      done();
-    });
-  });
-
-  it('refuses to add fields with uncreatable relation target class', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Relation', targetClass: 'not a valid class name'},
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
-      expect(error.error).toEqual('Invalid classname: not a valid class name, classnames can only have alphanumeric characters and _, and must start with an alpha character ');
-      done();
-    });
-  });
-
-  it('refuses to add fields with unknown types', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      foo: {type: 'Unknown'},
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INCORRECT_TYPE);
-      expect(error.error).toEqual('invalid field type: Unknown');
-      done();
-    });
-  });
-
-  it('will create classes', done => {
-    config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      aNumber: {type: 'Number'},
-      aString: {type: 'String'},
-      aBool: {type: 'Boolean'},
-      aDate: {type: 'Date'},
-      aObject: {type: 'Object'},
-      aArray: {type: 'Array'},
-      aGeoPoint: {type: 'GeoPoint'},
-      aFile: {type: 'File'},
-      aPointer: {type: 'Pointer', targetClass: 'ThisClassDoesNotExistYet'},
-      aRelation: {type: 'Relation', targetClass: 'NewClass'},
-      aBytes: {type: 'Bytes'},
-    }))
-    .then(actualSchema => {
-      const expectedSchema = {
-        className: 'NewClass',
-        fields: {
-          objectId: { type: 'String' },
-          updatedAt: { type: 'Date' },
-          createdAt: { type: 'Date' },
-          ACL: { type: 'ACL' },
-          aString: { type: 'String' },
-          aNumber: { type: 'Number' },
-          aBool: { type: 'Boolean' },
-          aDate: { type: 'Date' },
-          aObject: { type: 'Object' },
-          aArray: { type: 'Array' },
-          aGeoPoint: { type: 'GeoPoint' },
-          aFile: { type: 'File' },
-          aPointer: { type: 'Pointer', targetClass: 'ThisClassDoesNotExistYet' },
-          aRelation: { type: 'Relation', targetClass: 'NewClass' },
-          aBytes: {type: 'Bytes'},
-        },
-        classLevelPermissions: {
-          find: { '*': true },
-          get: { '*': true },
-          create: { '*': true },
-          update: { '*': true },
-          delete: { '*': true },
-          addField: { '*': true },
-        },
-      }
-      expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      done();
-    });
   });
 
   it('creates the default fields for non-custom classes', done => {
     config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('_Installation', {
-      foo: {type: 'Number'},
-    }))
-    .then(actualSchema => {
-      const expectedSchema = {
-        className: '_Installation',
-        fields: {
-          objectId: { type: 'String' },
-          updatedAt: { type: 'Date' },
-          createdAt: { type: 'Date' },
-          ACL: { type: 'ACL' },
-          foo: { type: 'Number' },
-          installationId: { type: 'String' },
-          deviceToken: { type: 'String' },
-          channels: { type: 'Array' },
-          deviceType: { type: 'String' },
-          pushType: { type: 'String' },
-          GCMSenderId: { type: 'String' },
-          timeZone: { type: 'String' },
-          localeIdentifier: { type: 'String' },
-          badge: { type: 'Number' },
-          appVersion: { type: 'String' },
-          appName: { type: 'String' },
-          appIdentifier: { type: 'String' },
-          parseVersion: { type: 'String' },
-        },
-        classLevelPermissions: {
-          find: { '*': true },
-          get: { '*': true },
-          create: { '*': true },
-          update: { '*': true },
-          delete: { '*': true },
-          addField: { '*': true },
-        },
-      }
-      expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      done();
-    });
+      .then(schema => schema.addClassIfNotExists('_Installation', {
+        foo: {type: 'Number'},
+      }))
+      .then(actualSchema => {
+        const expectedSchema = {
+          className: '_Installation',
+          fields: {
+            objectId: { type: 'String' },
+            updatedAt: { type: 'Date' },
+            createdAt: { type: 'Date' },
+            ACL: { type: 'ACL' },
+            foo: { type: 'Number' },
+            installationId: { type: 'String' },
+            deviceToken: { type: 'String' },
+            channels: { type: 'Array' },
+            deviceType: { type: 'String' },
+            pushType: { type: 'String' },
+            GCMSenderId: { type: 'String' },
+            timeZone: { type: 'String' },
+            localeIdentifier: { type: 'String' },
+            badge: { type: 'Number' },
+            appVersion: { type: 'String' },
+            appName: { type: 'String' },
+            appIdentifier: { type: 'String' },
+            parseVersion: { type: 'String' },
+          },
+          classLevelPermissions: {
+            find: { '*': true },
+            get: { '*': true },
+            create: { '*': true },
+            update: { '*': true },
+            delete: { '*': true },
+            addField: { '*': true },
+          },
+        }
+        expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
+        done();
+      });
   });
 
   it('creates non-custom classes which include relation field', done => {
     config.database.loadSchema()
     //as `_Role` is always created by default, we only get it here
-    .then(schema => schema.getOneSchema('_Role'))
-    .then(actualSchema => {
-      const expectedSchema = {
-        className: '_Role',
-        fields: {
-          objectId: { type: 'String' },
-          updatedAt: { type: 'Date' },
-          createdAt: { type: 'Date' },
-          ACL: { type: 'ACL' },
-          name: { type: 'String' },
-          users: { type: 'Relation', targetClass: '_User' },
-          roles: { type: 'Relation', targetClass: '_Role' },
-        },
-        classLevelPermissions: {
-          find: { '*': true },
-          get: { '*': true },
-          create: { '*': true },
-          update: { '*': true },
-          delete: { '*': true },
-          addField: { '*': true },
-        },
-      };
-      expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      done();
-    });
+      .then(schema => schema.getOneSchema('_Role'))
+      .then(actualSchema => {
+        const expectedSchema = {
+          className: '_Role',
+          fields: {
+            objectId: { type: 'String' },
+            updatedAt: { type: 'Date' },
+            createdAt: { type: 'Date' },
+            ACL: { type: 'ACL' },
+            name: { type: 'String' },
+            users: { type: 'Relation', targetClass: '_User' },
+            roles: { type: 'Relation', targetClass: '_Role' },
+          },
+          classLevelPermissions: {
+            find: { '*': true },
+            get: { '*': true },
+            create: { '*': true },
+            update: { '*': true },
+            delete: { '*': true },
+            addField: { '*': true },
+          },
+        };
+        expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
+        done();
+      });
   });
 
   it('creates non-custom classes which include pointer field', done => {
     config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('_Session', {}))
-    .then(actualSchema => {
-      const expectedSchema = {
-        className: '_Session',
-        fields: {
-          objectId: { type: 'String' },
-          updatedAt: { type: 'Date' },
-          createdAt: { type: 'Date' },
-          restricted: { type: 'Boolean' },
-          user: { type: 'Pointer', targetClass: '_User' },
-          installationId: { type: 'String' },
-          sessionToken: { type: 'String' },
-          expiresAt: { type: 'Date' },
-          createdWith: { type: 'Object' },
-          ACL: { type: 'ACL' },
-        },
-        classLevelPermissions: {
-          find: { '*': true },
-          get: { '*': true },
-          create: { '*': true },
-          update: { '*': true },
-          delete: { '*': true },
-          addField: { '*': true },
-        },
-      };
-      expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      done();
-    });
+      .then(schema => schema.addClassIfNotExists('_Session', {}))
+      .then(actualSchema => {
+        const expectedSchema = {
+          className: '_Session',
+          fields: {
+            objectId: { type: 'String' },
+            updatedAt: { type: 'Date' },
+            createdAt: { type: 'Date' },
+            restricted: { type: 'Boolean' },
+            user: { type: 'Pointer', targetClass: '_User' },
+            installationId: { type: 'String' },
+            sessionToken: { type: 'String' },
+            expiresAt: { type: 'Date' },
+            createdWith: { type: 'Object' },
+            ACL: { type: 'ACL' },
+          },
+          classLevelPermissions: {
+            find: { '*': true },
+            get: { '*': true },
+            create: { '*': true },
+            update: { '*': true },
+            delete: { '*': true },
+            addField: { '*': true },
+          },
+        };
+        expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
+        done();
+      });
   });
 
   it('refuses to create two geopoints', done => {
     config.database.loadSchema()
-    .then(schema => schema.addClassIfNotExists('NewClass', {
-      geo1: {type: 'GeoPoint'},
-      geo2: {type: 'GeoPoint'}
-    }))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INCORRECT_TYPE);
-      expect(error.error).toEqual('currently, only one GeoPoint field may exist in an object. Adding geo2 when geo1 already exists.');
-      done();
-    });
+      .then(schema => schema.addClassIfNotExists('NewClass', {
+        geo1: {type: 'GeoPoint'},
+        geo2: {type: 'GeoPoint'}
+      }))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INCORRECT_TYPE);
+        expect(error.error).toEqual('currently, only one GeoPoint field may exist in an object. Adding geo2 when geo1 already exists.');
+        done();
+      });
   });
 
   it('can check if a class exists', done => {
     config.database.loadSchema()
-    .then(schema => {
-      return schema.addClassIfNotExists('NewClass', {})
-      .then(() => {
-        schema.hasClass('NewClass')
-        .then(hasClass => {
-          expect(hasClass).toEqual(true);
-          done();
-        })
-        .catch(fail);
+      .then(schema => {
+        return schema.addClassIfNotExists('NewClass', {})
+          .then(() => {
+            schema.hasClass('NewClass')
+              .then(hasClass => {
+                expect(hasClass).toEqual(true);
+                done();
+              })
+              .catch(fail);
 
-        schema.hasClass('NonexistantClass')
-        .then(hasClass => {
-          expect(hasClass).toEqual(false);
-          done();
-        })
-        .catch(fail);
+            schema.hasClass('NonexistantClass')
+              .then(hasClass => {
+                expect(hasClass).toEqual(false);
+                done();
+              })
+              .catch(fail);
+          })
+          .catch(error => {
+            fail('Couldn\'t create class');
+            jfail(error);
+          });
       })
-      .catch(error => {
-        fail('Couldn\'t create class');
-        jfail(error);
-      });
-    })
-    .catch(() => fail('Couldn\'t load schema'));
+      .catch(() => fail('Couldn\'t load schema'));
   });
 
   it('refuses to delete fields from invalid class names', done => {
     config.database.loadSchema()
-    .then(schema => schema.deleteField('fieldName', 'invalid class name'))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
-      done();
-    });
+      .then(schema => schema.deleteField('fieldName', 'invalid class name'))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
+        done();
+      });
   });
 
   it('refuses to delete invalid fields', done => {
     config.database.loadSchema()
-    .then(schema => schema.deleteField('invalid field name', 'ValidClassName'))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_KEY_NAME);
-      done();
-    });
+      .then(schema => schema.deleteField('invalid field name', 'ValidClassName'))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_KEY_NAME);
+        done();
+      });
   });
 
   it('refuses to delete the default fields', done => {
     config.database.loadSchema()
-    .then(schema => schema.deleteField('installationId', '_Installation'))
-    .catch(error => {
-      expect(error.code).toEqual(136);
-      expect(error.message).toEqual('field installationId cannot be changed');
-      done();
-    });
+      .then(schema => schema.deleteField('installationId', '_Installation'))
+      .catch(error => {
+        expect(error.code).toEqual(136);
+        expect(error.message).toEqual('field installationId cannot be changed');
+        done();
+      });
   });
 
   it('refuses to delete fields from nonexistant classes', done => {
     config.database.loadSchema()
-    .then(schema => schema.deleteField('field', 'NoClass'))
-    .catch(error => {
-      expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
-      expect(error.message).toEqual('Class NoClass does not exist.');
-      done();
-    });
+      .then(schema => schema.deleteField('field', 'NoClass'))
+      .catch(error => {
+        expect(error.code).toEqual(Parse.Error.INVALID_CLASS_NAME);
+        expect(error.message).toEqual('Class NoClass does not exist.');
+        done();
+      });
   });
 
   it('refuses to delete fields that dont exist', done => {
     hasAllPODobject().save()
-    .then(() => config.database.loadSchema())
-    .then(schema => schema.deleteField('missingField', 'HasAllPOD'))
-    .fail(error => {
-      expect(error.code).toEqual(255);
-      expect(error.message).toEqual('Field missingField does not exist, cannot delete.');
-      done();
-    });
+      .then(() => config.database.loadSchema())
+      .then(schema => schema.deleteField('missingField', 'HasAllPOD'))
+      .fail(error => {
+        expect(error.code).toEqual(255);
+        expect(error.message).toEqual('Field missingField does not exist, cannot delete.');
+        done();
+      });
   });
 
   it('drops related collection when deleting relation field', done => {
@@ -785,54 +787,54 @@ describe('SchemaController', () => {
 
   it('can delete relation field when related _Join collection not exist', done => {
     config.database.loadSchema()
-    .then(schema => {
-      schema.addClassIfNotExists('NewClass', {
-        relationField: {type: 'Relation', targetClass: '_User'}
-      })
-      .then(actualSchema => {
-        const expectedSchema = {
-          className: 'NewClass',
-          fields: {
-            objectId: { type: 'String' },
-            updatedAt: { type: 'Date' },
-            createdAt: { type: 'Date' },
-            ACL: { type: 'ACL' },
-            relationField: { type: 'Relation', targetClass: '_User' },
-          },
-          classLevelPermissions: {
-            find: { '*': true },
-            get: { '*': true },
-            create: { '*': true },
-            update: { '*': true },
-            delete: { '*': true },
-            addField: { '*': true },
-          },
-        };
-        expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
-      })
-      .then(() => config.database.collectionExists('_Join:relationField:NewClass'))
-      .then(exist => {
-        on_db('postgres', () => {
-          // We create the table when creating the column
-          expect(exist).toEqual(true);
-        }, () => {
-          expect(exist).toEqual(false);
-        });
+      .then(schema => {
+        schema.addClassIfNotExists('NewClass', {
+          relationField: {type: 'Relation', targetClass: '_User'}
+        })
+          .then(actualSchema => {
+            const expectedSchema = {
+              className: 'NewClass',
+              fields: {
+                objectId: { type: 'String' },
+                updatedAt: { type: 'Date' },
+                createdAt: { type: 'Date' },
+                ACL: { type: 'ACL' },
+                relationField: { type: 'Relation', targetClass: '_User' },
+              },
+              classLevelPermissions: {
+                find: { '*': true },
+                get: { '*': true },
+                create: { '*': true },
+                update: { '*': true },
+                delete: { '*': true },
+                addField: { '*': true },
+              },
+            };
+            expect(dd(actualSchema, expectedSchema)).toEqual(undefined);
+          })
+          .then(() => config.database.collectionExists('_Join:relationField:NewClass'))
+          .then(exist => {
+            on_db('postgres', () => {
+              // We create the table when creating the column
+              expect(exist).toEqual(true);
+            }, () => {
+              expect(exist).toEqual(false);
+            });
 
-      })
-      .then(() => schema.deleteField('relationField', 'NewClass', config.database))
-      .then(() => schema.reloadData())
-      .then(() => {
-        const expectedSchema = {
-          objectId: { type: 'String' },
-          updatedAt: { type: 'Date' },
-          createdAt: { type: 'Date' },
-          ACL: { type: 'ACL' },
-        };
-        expect(dd(schema.data.NewClass, expectedSchema)).toEqual(undefined);
-        done();
+          })
+          .then(() => schema.deleteField('relationField', 'NewClass', config.database))
+          .then(() => schema.reloadData())
+          .then(() => {
+            const expectedSchema = {
+              objectId: { type: 'String' },
+              updatedAt: { type: 'Date' },
+              createdAt: { type: 'Date' },
+              ACL: { type: 'ACL' },
+            };
+            expect(dd(schema.data.NewClass, expectedSchema)).toEqual(undefined);
+            done();
+          });
       });
-    });
   });
 
   it('can delete string fields and resave as number field', done => {
@@ -840,53 +842,53 @@ describe('SchemaController', () => {
     var obj1 = hasAllPODobject();
     var obj2 = hasAllPODobject();
     Parse.Object.saveAll([obj1, obj2])
-    .then(() => config.database.loadSchema())
-    .then(schema => schema.deleteField('aString', 'HasAllPOD', config.database))
-    .then(() => new Parse.Query('HasAllPOD').get(obj1.id))
-    .then(obj1Reloaded => {
-      expect(obj1Reloaded.get('aString')).toEqual(undefined);
-      obj1Reloaded.set('aString', ['not a string', 'this time']);
-      obj1Reloaded.save()
-      .then(obj1reloadedAgain => {
-        expect(obj1reloadedAgain.get('aString')).toEqual(['not a string', 'this time']);
-        return new Parse.Query('HasAllPOD').get(obj2.id);
+      .then(() => config.database.loadSchema())
+      .then(schema => schema.deleteField('aString', 'HasAllPOD', config.database))
+      .then(() => new Parse.Query('HasAllPOD').get(obj1.id))
+      .then(obj1Reloaded => {
+        expect(obj1Reloaded.get('aString')).toEqual(undefined);
+        obj1Reloaded.set('aString', ['not a string', 'this time']);
+        obj1Reloaded.save()
+          .then(obj1reloadedAgain => {
+            expect(obj1reloadedAgain.get('aString')).toEqual(['not a string', 'this time']);
+            return new Parse.Query('HasAllPOD').get(obj2.id);
+          })
+          .then(obj2reloaded => {
+            expect(obj2reloaded.get('aString')).toEqual(undefined);
+            done();
+            Parse.Object.enableSingleInstance();
+          });
       })
-      .then(obj2reloaded => {
-        expect(obj2reloaded.get('aString')).toEqual(undefined);
+      .catch(error => {
+        jfail(error);
         done();
-        Parse.Object.enableSingleInstance();
       });
-    })
-    .catch(error => {
-      jfail(error);
-      done();
-    });
   });
 
   it('can delete pointer fields and resave as string', done => {
     Parse.Object.disableSingleInstance();
     var obj1 = new Parse.Object('NewClass');
     obj1.save()
-    .then(() => {
-      obj1.set('aPointer', obj1);
-      return obj1.save();
-    })
-    .then(obj1 => {
-      expect(obj1.get('aPointer').id).toEqual(obj1.id);
-    })
-    .then(() => config.database.loadSchema())
-    .then(schema => schema.deleteField('aPointer', 'NewClass', config.database))
-    .then(() => new Parse.Query('NewClass').get(obj1.id))
-    .then(obj1 => {
-      expect(obj1.get('aPointer')).toEqual(undefined);
-      obj1.set('aPointer', 'Now a string');
-      return obj1.save();
-    })
-    .then(obj1 => {
-      expect(obj1.get('aPointer')).toEqual('Now a string');
-      done();
-      Parse.Object.enableSingleInstance();
-    });
+      .then(() => {
+        obj1.set('aPointer', obj1);
+        return obj1.save();
+      })
+      .then(obj1 => {
+        expect(obj1.get('aPointer').id).toEqual(obj1.id);
+      })
+      .then(() => config.database.loadSchema())
+      .then(schema => schema.deleteField('aPointer', 'NewClass', config.database))
+      .then(() => new Parse.Query('NewClass').get(obj1.id))
+      .then(obj1 => {
+        expect(obj1.get('aPointer')).toEqual(undefined);
+        obj1.set('aPointer', 'Now a string');
+        return obj1.save();
+      })
+      .then(obj1 => {
+        expect(obj1.get('aPointer')).toEqual('Now a string');
+        done();
+        Parse.Object.enableSingleInstance();
+      });
   });
 
   it('can merge schemas', done => {
