@@ -17,17 +17,17 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       },
       publicServerURL: "https://my.public.server.com/1"
     })
-    .then(() => {
-      var config = new Config("test");
-      expect(config.invalidLinkURL).toEqual("myInvalidLink");
-      expect(config.verifyEmailSuccessURL).toEqual("myVerifyEmailSuccess");
-      expect(config.choosePasswordURL).toEqual("myChoosePassword");
-      expect(config.passwordResetSuccessURL).toEqual("myPasswordResetSuccess");
-      expect(config.parseFrameURL).toEqual("http://example.com/handle-parse-iframe");
-      expect(config.verifyEmailURL).toEqual("https://my.public.server.com/1/apps/test/verify_email");
-      expect(config.requestResetPasswordURL).toEqual("https://my.public.server.com/1/apps/test/request_password_reset");
-      done();
-    });
+      .then(() => {
+        var config = new Config("test");
+        expect(config.invalidLinkURL).toEqual("myInvalidLink");
+        expect(config.verifyEmailSuccessURL).toEqual("myVerifyEmailSuccess");
+        expect(config.choosePasswordURL).toEqual("myChoosePassword");
+        expect(config.passwordResetSuccessURL).toEqual("myPasswordResetSuccess");
+        expect(config.parseFrameURL).toEqual("http://example.com/handle-parse-iframe");
+        expect(config.verifyEmailURL).toEqual("https://my.public.server.com/1/apps/test/verify_email");
+        expect(config.requestResetPasswordURL).toEqual("https://my.public.server.com/1/apps/test/request_password_reset");
+        done();
+      });
   });
 
   it('sends verification email if email verification is enabled', done => {
@@ -42,27 +42,27 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      spyOn(emailAdapter, 'sendVerificationEmail');
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.setEmail('testIfEnabled@parse.com');
-      user.signUp(null, {
-        success: function(user) {
-          expect(emailAdapter.sendVerificationEmail).toHaveBeenCalled();
-          user.fetch()
-          .then(() => {
-            expect(user.get('emailVerified')).toEqual(false);
+      .then(() => {
+        spyOn(emailAdapter, 'sendVerificationEmail');
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.setEmail('testIfEnabled@parse.com');
+        user.signUp(null, {
+          success: function(user) {
+            expect(emailAdapter.sendVerificationEmail).toHaveBeenCalled();
+            user.fetch()
+              .then(() => {
+                expect(user.get('emailVerified')).toEqual(false);
+                done();
+              });
+          },
+          error: function() {
+            fail('Failed to save user');
             done();
-          });
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+          }
+        });
       });
-    });
   });
 
   it('does not send verification email when verification is enabled and email is not set', done => {
@@ -77,26 +77,26 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      spyOn(emailAdapter, 'sendVerificationEmail');
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.signUp(null, {
-        success: function(user) {
-          expect(emailAdapter.sendVerificationEmail).not.toHaveBeenCalled();
-          user.fetch()
-          .then(() => {
-            expect(user.get('emailVerified')).toEqual(undefined);
+      .then(() => {
+        spyOn(emailAdapter, 'sendVerificationEmail');
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.signUp(null, {
+          success: function(user) {
+            expect(emailAdapter.sendVerificationEmail).not.toHaveBeenCalled();
+            user.fetch()
+              .then(() => {
+                expect(user.get('emailVerified')).toEqual(undefined);
+                done();
+              });
+          },
+          error: function() {
+            fail('Failed to save user');
             done();
-          });
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+          }
+        });
       });
-    });
   });
 
   it('does send a validation email when updating the email', done => {
@@ -111,35 +111,35 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      spyOn(emailAdapter, 'sendVerificationEmail');
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.signUp(null, {
-        success: function(user) {
-          expect(emailAdapter.sendVerificationEmail).not.toHaveBeenCalled();
-          user.fetch()
-          .then((user) => {
-            user.set("email", "testWhenUpdating@parse.com");
-            return user.save();
-          }).then((user) => {
-            return user.fetch();
-          }).then(() => {
-            expect(user.get('emailVerified')).toEqual(false);
-            // Wait as on update email, we need to fetch the username
-            setTimeout(function(){
-              expect(emailAdapter.sendVerificationEmail).toHaveBeenCalled();
-              done();
-            }, 200);
-          });
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+      .then(() => {
+        spyOn(emailAdapter, 'sendVerificationEmail');
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.signUp(null, {
+          success: function(user) {
+            expect(emailAdapter.sendVerificationEmail).not.toHaveBeenCalled();
+            user.fetch()
+              .then((user) => {
+                user.set("email", "testWhenUpdating@parse.com");
+                return user.save();
+              }).then((user) => {
+                return user.fetch();
+              }).then(() => {
+                expect(user.get('emailVerified')).toEqual(false);
+                // Wait as on update email, we need to fetch the username
+                setTimeout(function(){
+                  expect(emailAdapter.sendVerificationEmail).toHaveBeenCalled();
+                  done();
+                }, 200);
+              });
+          },
+          error: function() {
+            fail('Failed to save user');
+            done();
+          }
+        });
       });
-    });
   });
 
   it('does send a validation email with valid verification link when updating the email', done => {
@@ -154,39 +154,39 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      spyOn(emailAdapter, 'sendVerificationEmail').and.callFake((options) => {
-        expect(options.link).not.toBeNull();
-        expect(options.link).not.toMatch(/token=undefined/);
-        Promise.resolve();
+      .then(() => {
+        spyOn(emailAdapter, 'sendVerificationEmail').and.callFake((options) => {
+          expect(options.link).not.toBeNull();
+          expect(options.link).not.toMatch(/token=undefined/);
+          Promise.resolve();
+        });
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.signUp(null, {
+          success: function(user) {
+            expect(emailAdapter.sendVerificationEmail).not.toHaveBeenCalled();
+            user.fetch()
+              .then((user) => {
+                user.set("email", "testValidLinkWhenUpdating@parse.com");
+                return user.save();
+              }).then((user) => {
+                return user.fetch();
+              }).then(() => {
+                expect(user.get('emailVerified')).toEqual(false);
+                // Wait as on update email, we need to fetch the username
+                setTimeout(function(){
+                  expect(emailAdapter.sendVerificationEmail).toHaveBeenCalled();
+                  done();
+                }, 200);
+              });
+          },
+          error: function() {
+            fail('Failed to save user');
+            done();
+          }
+        });
       });
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.signUp(null, {
-        success: function(user) {
-          expect(emailAdapter.sendVerificationEmail).not.toHaveBeenCalled();
-          user.fetch()
-          .then((user) => {
-            user.set("email", "testValidLinkWhenUpdating@parse.com");
-            return user.save();
-          }).then((user) => {
-            return user.fetch();
-          }).then(() => {
-            expect(user.get('emailVerified')).toEqual(false);
-            // Wait as on update email, we need to fetch the username
-            setTimeout(function(){
-              expect(emailAdapter.sendVerificationEmail).toHaveBeenCalled();
-              done();
-            }, 200);
-          });
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
-      });
-    });
   });
 
   it('does send with a simple adapter', done => {
@@ -211,33 +211,33 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testSendSimpleAdapter@parse.com");
-      user.signUp(null, {
-        success: function(user) {
-          expect(calls).toBe(1);
-          user.fetch()
-          .then((user) => {
-            return user.save();
-          }).then(() => {
-            return Parse.User.requestPasswordReset("testSendSimpleAdapter@parse.com").catch(() => {
-              fail('Should not fail requesting a password');
-              done();
-            })
-          }).then(() => {
-            expect(calls).toBe(2);
+      .then(() => {
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testSendSimpleAdapter@parse.com");
+        user.signUp(null, {
+          success: function(user) {
+            expect(calls).toBe(1);
+            user.fetch()
+              .then((user) => {
+                return user.save();
+              }).then(() => {
+                return Parse.User.requestPasswordReset("testSendSimpleAdapter@parse.com").catch(() => {
+                  fail('Should not fail requesting a password');
+                  done();
+                })
+              }).then(() => {
+                expect(calls).toBe(2);
+                done();
+              });
+          },
+          error: function() {
+            fail('Failed to save user');
             done();
-          });
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+          }
+        });
       });
-    });
   });
 
   it('prevents user from login if email is not verified but preventLoginWithUnverifiedEmail is set to true', done => {
@@ -252,25 +252,25 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
         domain: 'd',
       }),
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testInvalidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.logIn("zxcv", "asdf"))
       .then(() => {
-        fail('login should have failed');
-        done();
-      }, error => {
-        expect(error.message).toEqual('User email is not verified.')
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testInvalidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.logIn("zxcv", "asdf"))
+          .then(() => {
+            fail('login should have failed');
+            done();
+          }, error => {
+            expect(error.message).toEqual('User email is not verified.')
+            done();
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('allows user to login only after user clicks on the link to confirm email address if preventLoginWithUnverifiedEmail is set to true', done => {
@@ -290,41 +290,41 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setPassword("other-password");
-      user.setUsername("user");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
-      expect(sendEmailOptions).not.toBeUndefined();
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=user');
-        user.fetch()
-        .then(() => {
-          expect(user.get('emailVerified')).toEqual(true);
+      .then(() => {
+        user.setPassword("other-password");
+        user.setUsername("user");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
+        expect(sendEmailOptions).not.toBeUndefined();
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=user');
+          user.fetch()
+            .then(() => {
+              expect(user.get('emailVerified')).toEqual(true);
 
-          Parse.User.logIn("user", "other-password")
-          .then(user => {
-            expect(typeof user).toBe('object');
-            expect(user.get('emailVerified')).toBe(true);
-            done();
-          }, () => {
-            fail('login should have succeeded');
-            done();
-          });
-        }, (err) => {
-          jfail(err);
-          fail("this should not fail");
-          done();
-        }).catch((err) => {
-          jfail(err);
-          done();
-        })
+              Parse.User.logIn("user", "other-password")
+                .then(user => {
+                  expect(typeof user).toBe('object');
+                  expect(user.get('emailVerified')).toBe(true);
+                  done();
+                }, () => {
+                  fail('login should have succeeded');
+                  done();
+                });
+            }, (err) => {
+              jfail(err);
+              fail("this should not fail");
+              done();
+            }).catch((err) => {
+              jfail(err);
+              done();
+            })
+        });
       });
-    });
   });
 
   it('allows user to login if email is not verified but preventLoginWithUnverifiedEmail is set to false', done => {
@@ -339,26 +339,26 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
         domain: 'd',
       }),
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testInvalidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.logIn("zxcv", "asdf"))
-      .then(user => {
-        expect(typeof user).toBe('object');
-        expect(user.get('emailVerified')).toBe(false);
-        done();
-      }, () => {
-        fail('login should have succeeded');
+      .then(() => {
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testInvalidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.logIn("zxcv", "asdf"))
+          .then(user => {
+            expect(typeof user).toBe('object');
+            expect(user.get('emailVerified')).toBe(false);
+            done();
+          }, () => {
+            fail('login should have succeeded');
+            done();
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('fails if you include an emailAdapter, set a publicServerURL, but have no appName and send a password reset email', done => {
@@ -371,26 +371,26 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
         domain: 'd',
       }),
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testInvalidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
-      .then(result => {
-        console.log(result);
-        fail('sending password reset email should not have succeeded');
-        done();
-      }, error => {
-        expect(error.message).toEqual('An appName, publicServerURL, and emailAdapter are required for password reset and email verification functionality.')
+      .then(() => {
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testInvalidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
+          .then(result => {
+            console.log(result);
+            fail('sending password reset email should not have succeeded');
+            done();
+          }, error => {
+            expect(error.message).toEqual('An appName, publicServerURL, and emailAdapter are required for password reset and email verification functionality.')
+            done();
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('fails if you include an emailAdapter, have an appName, but have no publicServerURL and send a password reset email', done => {
@@ -402,26 +402,26 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
         domain: 'd',
       }),
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testInvalidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
-      .then(result => {
-        console.log(result);
-        fail('sending password reset email should not have succeeded');
-        done();
-      }, error => {
-        expect(error.message).toEqual('An appName, publicServerURL, and emailAdapter are required for password reset and email verification functionality.')
+      .then(() => {
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testInvalidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
+          .then(result => {
+            console.log(result);
+            fail('sending password reset email should not have succeeded');
+            done();
+          }, error => {
+            expect(error.message).toEqual('An appName, publicServerURL, and emailAdapter are required for password reset and email verification functionality.')
+            done();
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('fails if you set a publicServerURL, have an appName, but no emailAdapter and send a password reset email', done => {
@@ -430,26 +430,26 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       publicServerURL: 'http://localhost:1337/1',
       emailAdapter: undefined,
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testInvalidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
-      .then(result => {
-        console.log(result);
-        fail('sending password reset email should not have succeeded');
-        done();
-      }, error => {
-        expect(error.message).toEqual('An appName, publicServerURL, and emailAdapter are required for password reset and email verification functionality.')
+      .then(() => {
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testInvalidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
+          .then(result => {
+            console.log(result);
+            fail('sending password reset email should not have succeeded');
+            done();
+          }, error => {
+            expect(error.message).toEqual('An appName, publicServerURL, and emailAdapter are required for password reset and email verification functionality.')
+            done();
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('succeeds sending a password reset email if appName, publicServerURL, and email adapter are prodvided', done => {
@@ -462,23 +462,23 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
         domain: 'd',
       }),
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set("email", "testInvalidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
       .then(() => {
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set("email", "testInvalidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.requestPasswordReset("testInvalidConfig@parse.com"))
+          .then(() => {
+            done();
+          }, error => {
+            done(error);
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
-      }, error => {
-        done(error);
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('succeeds sending a password reset username if appName, publicServerURL, and email adapter are prodvided', done => {
@@ -501,23 +501,23 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       publicServerURL: 'http://localhost:1337/1',
       emailAdapter: adapter
     })
-    .then(() => {
-      const user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("testValidConfig@parse.com");
-      user.signUp(null)
-      .then(() => Parse.User.requestPasswordReset("testValidConfig@parse.com"))
       .then(() => {
-        expect(adapter.sendMail).toHaveBeenCalled();
+        const user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("testValidConfig@parse.com");
+        user.signUp(null)
+          .then(() => Parse.User.requestPasswordReset("testValidConfig@parse.com"))
+          .then(() => {
+            expect(adapter.sendMail).toHaveBeenCalled();
+            done();
+          }, error => {
+            done(error);
+          });
+      })
+      .catch(error => {
+        fail(JSON.stringify(error));
         done();
-      }, error => {
-        done(error);
       });
-    })
-    .catch(error => {
-      fail(JSON.stringify(error));
-      done();
-    });
   });
 
   it('does not send verification email if email verification is disabled', done => {
@@ -532,26 +532,26 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       verifyUserEmails: false,
       emailAdapter: emailAdapter,
     })
-    .then(() => {
-      spyOn(emailAdapter, 'sendVerificationEmail');
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.signUp(null, {
-        success: function(user) {
-          user.fetch()
-          .then(() => {
-            expect(emailAdapter.sendVerificationEmail.calls.count()).toEqual(0);
-            expect(user.get('emailVerified')).toEqual(undefined);
+      .then(() => {
+        spyOn(emailAdapter, 'sendVerificationEmail');
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.signUp(null, {
+          success: function(user) {
+            user.fetch()
+              .then(() => {
+                expect(emailAdapter.sendVerificationEmail.calls.count()).toEqual(0);
+                expect(user.get('emailVerified')).toEqual(undefined);
+                done();
+              });
+          },
+          error: function() {
+            fail('Failed to save user');
             done();
-          });
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+          }
+        });
       });
-    });
   });
 
   it('receives the app name and user in the adapter', done => {
@@ -571,22 +571,22 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      var user = new Parse.User();
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set('email', 'user@parse.com');
-      user.signUp(null, {
-        success: () => {
-          expect(emailSent).toBe(true);
-          done();
-        },
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+      .then(() => {
+        var user = new Parse.User();
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set('email', 'user@parse.com');
+        user.signUp(null, {
+          success: () => {
+            expect(emailSent).toBe(true);
+            done();
+          },
+          error: function() {
+            fail('Failed to save user');
+            done();
+          }
+        });
       });
-    });
   })
 
   it('when you click the link in the email it sets emailVerified to true and redirects you', done => {
@@ -605,32 +605,32 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setPassword("other-password");
-      user.setUsername("user");
-      user.set('email', 'user@parse.com');
-      return user.signUp();
-    }).then(() => {
-      expect(sendEmailOptions).not.toBeUndefined();
-      request.get(sendEmailOptions.link, {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=user');
-        user.fetch()
-        .then(() => {
-          expect(user.get('emailVerified')).toEqual(true);
-          done();
-        }, (err) => {
-          jfail(err);
-          fail("this should not fail");
-          done();
-        }).catch((err) => {
-          jfail(err);
-          done();
-        })
+      .then(() => {
+        user.setPassword("other-password");
+        user.setUsername("user");
+        user.set('email', 'user@parse.com');
+        return user.signUp();
+      }).then(() => {
+        expect(sendEmailOptions).not.toBeUndefined();
+        request.get(sendEmailOptions.link, {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/verify_email_success.html?username=user');
+          user.fetch()
+            .then(() => {
+              expect(user.get('emailVerified')).toEqual(true);
+              done();
+            }, (err) => {
+              jfail(err);
+              fail("this should not fail");
+              done();
+            }).catch((err) => {
+              jfail(err);
+              done();
+            })
+        });
       });
-    });
   });
 
   it('redirects you to invalid link if you try to verify email incorrecly', done => {
@@ -644,15 +644,15 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      request.get('http://localhost:8378/1/apps/test/verify_email', {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_link.html');
-        done()
+      .then(() => {
+        request.get('http://localhost:8378/1/apps/test/verify_email', {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_link.html');
+          done()
+        });
       });
-    });
   });
 
   it('redirects you to invalid verification link page if you try to validate a nonexistant users email', done => {
@@ -666,15 +666,15 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      request.get('http://localhost:8378/1/apps/test/verify_email?token=asdfasdf&username=sadfasga', {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=sadfasga&appId=test');
-        done();
+      .then(() => {
+        request.get('http://localhost:8378/1/apps/test/verify_email?token=asdfasdf&username=sadfasga', {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=sadfasga&appId=test');
+          done();
+        });
       });
-    });
   });
 
   it('redirects you to link send fail page if you try to resend a link for a nonexistant user', done => {
@@ -688,18 +688,18 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      request.post('http://localhost:8378/1/apps/test/resend_verification_email', {
-        followRedirect: false,
-        form: {
-          username: "sadfasga"
-        }
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/link_send_fail.html');
-        done();
+      .then(() => {
+        request.post('http://localhost:8378/1/apps/test/resend_verification_email', {
+          followRedirect: false,
+          form: {
+            username: "sadfasga"
+          }
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/link_send_fail.html');
+          done();
+        });
       });
-    });
   });
 
   it('does not update email verified if you use an invalid token', done => {
@@ -712,10 +712,10 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
           expect(response.statusCode).toEqual(302);
           expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_verification_link.html?username=zxcv&appId=test');
           user.fetch()
-          .then(() => {
-            expect(user.get('emailVerified')).toEqual(false);
-            done();
-          });
+            .then(() => {
+              expect(user.get('emailVerified')).toEqual(false);
+              done();
+            });
         });
       },
       sendPasswordResetEmail: () => Promise.resolve(),
@@ -727,18 +727,18 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set('email', 'user@parse.com');
-      user.signUp(null, {
-        success: () => {},
-        error: function() {
-          fail('Failed to save user');
-          done();
-        }
+      .then(() => {
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set('email', 'user@parse.com');
+        user.signUp(null, {
+          success: () => {},
+          error: function() {
+            fail('Failed to save user');
+            done();
+          }
+        });
       });
-    });
   });
 
   it('should send a password reset link', done => {
@@ -768,20 +768,20 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setPassword("asdf");
-      user.setUsername("zxcv+zxcv");
-      user.set('email', 'user@parse.com');
-      user.signUp().then(() => {
-        Parse.User.requestPasswordReset('user@parse.com', {
-          error: (err) => {
-            jfail(err);
-            fail("Should not fail requesting a password");
-            done();
-          }
+      .then(() => {
+        user.setPassword("asdf");
+        user.setUsername("zxcv+zxcv");
+        user.set('email', 'user@parse.com');
+        user.signUp().then(() => {
+          Parse.User.requestPasswordReset('user@parse.com', {
+            error: (err) => {
+              jfail(err);
+              fail("Should not fail requesting a password");
+              done();
+            }
+          });
         });
       });
-    });
   });
 
   it('redirects you to invalid link if you try to request password for a nonexistant users email', done => {
@@ -795,15 +795,15 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       },
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      request.get('http://localhost:8378/1/apps/test/request_password_reset?token=asdfasdf&username=sadfasga', {
-        followRedirect: false,
-      }, (error, response) => {
-        expect(response.statusCode).toEqual(302);
-        expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_link.html');
-        done();
+      .then(() => {
+        request.get('http://localhost:8378/1/apps/test/request_password_reset?token=asdfasdf&username=sadfasga', {
+          followRedirect: false,
+        }, (error, response) => {
+          expect(response.statusCode).toEqual(302);
+          expect(response.body).toEqual('Found. Redirecting to http://localhost:8378/1/apps/invalid_link.html');
+          done();
+        });
       });
-    });
   });
 
   it('should programatically reset password', done => {
@@ -848,12 +848,12 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
             Parse.User.logIn("zxcv", "hello").then(function(){
               const config = new Config('test');
               config.database.adapter.find('_User', { fields: {} }, { 'username': 'zxcv' }, { limit: 1 })
-              .then(results => {
+                .then(results => {
                 // _perishable_token should be unset after reset password
-                expect(results.length).toEqual(1);
-                expect(results[0]['_perishable_token']).toEqual(undefined);
-                done();
-              });
+                  expect(results.length).toEqual(1);
+                  expect(results[0]['_perishable_token']).toEqual(undefined);
+                  done();
+                });
             }, (err) => {
               jfail(err);
               fail("should login with new password");
@@ -871,19 +871,19 @@ describe("Custom Pages, Email Verification, Password Reset", () => {
       emailAdapter: emailAdapter,
       publicServerURL: "http://localhost:8378/1"
     })
-    .then(() => {
-      user.setPassword("asdf");
-      user.setUsername("zxcv");
-      user.set('email', 'user@parse.com');
-      user.signUp().then(() => {
-        Parse.User.requestPasswordReset('user@parse.com', {
-          error: (err) => {
-            jfail(err);
-            fail("Should not fail");
-            done();
-          }
+      .then(() => {
+        user.setPassword("asdf");
+        user.setUsername("zxcv");
+        user.set('email', 'user@parse.com');
+        user.signUp().then(() => {
+          Parse.User.requestPasswordReset('user@parse.com', {
+            error: (err) => {
+              jfail(err);
+              fail("Should not fail");
+              done();
+            }
+          });
         });
       });
-    });
   });
 })
