@@ -18,8 +18,8 @@ describe('info logs', () => {
         } else {
           expect(results[0].message).toEqual('testing info logs');
         }
-          // Check the error log
-          // Regression #2639
+        // Check the error log
+        // Regression #2639
         winstonLoggerAdapter.query({
           from: new Date(Date.now() - 500),
           size: 100,
@@ -58,42 +58,42 @@ describe('error logs', () => {
 describe('verbose logs', () => {
   it("mask sensitive information in _User class", (done) => {
     reconfigureServer({ verbose: true })
-    .then(() => createTestUser())
-    .then(() => {
-      const winstonLoggerAdapter = new WinstonLoggerAdapter();
-      return winstonLoggerAdapter.query({
-        from: new Date(Date.now() - 500),
-        size: 100,
-        level: 'verbose'
-      });
-    }).then((results) => {
-      const logString = JSON.stringify(results);
-      expect(logString.match(/\*\*\*\*\*\*\*\*/g).length).not.toBe(0);
-      expect(logString.match(/moon-y/g)).toBe(null);
-
-      var headers = {
-        'X-Parse-Application-Id': 'test',
-        'X-Parse-REST-API-Key': 'rest'
-      };
-      request.get({
-        headers: headers,
-        url: 'http://localhost:8378/1/login?username=test&password=moon-y'
-      }, () => {
+      .then(() => createTestUser())
+      .then(() => {
         const winstonLoggerAdapter = new WinstonLoggerAdapter();
         return winstonLoggerAdapter.query({
           from: new Date(Date.now() - 500),
           size: 100,
           level: 'verbose'
-        }).then((results) => {
-          const logString = JSON.stringify(results);
-          expect(logString.match(/\*\*\*\*\*\*\*\*/g).length).not.toBe(0);
-          expect(logString.match(/moon-y/g)).toBe(null);
-          done();
         });
-      });
-    }).catch((err) => {
-      fail(JSON.stringify(err));
-      done();
-    })
+      }).then((results) => {
+        const logString = JSON.stringify(results);
+        expect(logString.match(/\*\*\*\*\*\*\*\*/g).length).not.toBe(0);
+        expect(logString.match(/moon-y/g)).toBe(null);
+
+        var headers = {
+          'X-Parse-Application-Id': 'test',
+          'X-Parse-REST-API-Key': 'rest'
+        };
+        request.get({
+          headers: headers,
+          url: 'http://localhost:8378/1/login?username=test&password=moon-y'
+        }, () => {
+          const winstonLoggerAdapter = new WinstonLoggerAdapter();
+          return winstonLoggerAdapter.query({
+            from: new Date(Date.now() - 500),
+            size: 100,
+            level: 'verbose'
+          }).then((results) => {
+            const logString = JSON.stringify(results);
+            expect(logString.match(/\*\*\*\*\*\*\*\*/g).length).not.toBe(0);
+            expect(logString.match(/moon-y/g)).toBe(null);
+            done();
+          });
+        });
+      }).catch((err) => {
+        fail(JSON.stringify(err));
+        done();
+      })
   });
 });
