@@ -9,7 +9,7 @@ var logger = require('../../logger').default;
 // Returns a promise that fulfills iff this user id is valid.
 function validateAuthData(authData, params) {
   return vkOAuth2Request(params).then(function (response) {
-    if (response && response && response.access_token) {
+    if (response && response.access_token) {
       return request("api.vk.com", "method/secure.checkToken?token=" + authData.access_token + "&client_secret=" + params.appSecret + "&access_token=" + response.access_token).then(function (response) {
         if (response && response.response && response.response.user_id == authData.id) {
           return;
@@ -23,14 +23,15 @@ function validateAuthData(authData, params) {
 }
 
 function vkOAuth2Request(params) {
-  var promise = new Parse.Promise();
-  return promise.then(function(){
+  return new Promise(function (resolve) {
     if (!params || !params.appIds || !params.appIds.length || !params.appSecret || !params.appSecret.length) {
       logger.error('Vk Auth', 'Vk auth is not configured. Missing appIds or appSecret.');
       throw new Parse.Error(Parse.Error.OBJECT_NOT_FOUND, 'Vk auth is not configured. Missing appIds or appSecret.');
     }
-    return request("oauth.vk.com", "access_token?client_id=" + params.appIds + "&client_secret=" + params.appSecret + "&v=5.59&grant_type=client_credentials")
-  })
+    resolve();
+  }).then(function () {
+    return request("oauth.vk.com", "access_token?client_id=" + params.appIds + "&client_secret=" + params.appSecret + "&v=5.59&grant_type=client_credentials");
+  });
 }
 
 // Returns a promise that fulfills iff this app id is valid.
