@@ -25,6 +25,7 @@ function mongoFieldToParseSchemaField(type) {
   case 'geopoint': return {type: 'GeoPoint'};
   case 'file':     return {type: 'File'};
   case 'bytes':    return {type: 'Bytes'};
+  case 'polygon':  return {type: 'Polygon'};
   }
 }
 
@@ -98,6 +99,7 @@ function parseFieldTypeToMongoFieldType({ type, targetClass }) {
   case 'GeoPoint': return 'geopoint';
   case 'File':     return 'file';
   case 'Bytes':    return 'bytes';
+  case 'Polygon':  return 'polygon';
   }
 }
 
@@ -110,10 +112,10 @@ class MongoSchemaCollection {
 
   _fetchAllSchemasFrom_SCHEMA() {
     return this._collection._rawFind({})
-    .then(schemas => schemas.map(mongoSchemaToParseSchema));
+      .then(schemas => schemas.map(mongoSchemaToParseSchema));
   }
 
-  _fechOneSchemaFrom_SCHEMA(name: string) {
+  _fetchOneSchemaFrom_SCHEMA(name: string) {
     return this._collection._rawFind(_mongoSchemaQueryFromNameQuery(name), { limit: 1 }).then(results => {
       if (results.length === 1) {
         return mongoSchemaToParseSchema(results[0]);
@@ -148,33 +150,33 @@ class MongoSchemaCollection {
 
   // TODO: don't spend an extra query on finding the schema if the type we are trying to add isn't a GeoPoint.
   addFieldIfNotExists(className: string, fieldName: string, type: string) {
-    return this._fechOneSchemaFrom_SCHEMA(className)
-    .then(schema => {
+    return this._fetchOneSchemaFrom_SCHEMA(className)
+      .then(schema => {
       // The schema exists. Check for existing GeoPoints.
-      if (type.type === 'GeoPoint') {
+        if (type.type === 'GeoPoint') {
         // Make sure there are not other geopoint fields
-        if (Object.keys(schema.fields).some(existingField => schema.fields[existingField].type === 'GeoPoint')) {
-          throw new Parse.Error(Parse.Error.INCORRECT_TYPE, 'MongoDB only supports one GeoPoint field in a class.');
+          if (Object.keys(schema.fields).some(existingField => schema.fields[existingField].type === 'GeoPoint')) {
+            throw new Parse.Error(Parse.Error.INCORRECT_TYPE, 'MongoDB only supports one GeoPoint field in a class.');
+          }
         }
-      }
-      return;
-    }, error => {
+        return;
+      }, error => {
       // If error is undefined, the schema doesn't exist, and we can create the schema with the field.
       // If some other error, reject with it.
-      if (error === undefined) {
-        return;
-      }
-      throw error;
-    })
-    .then(() => {
+        if (error === undefined) {
+          return;
+        }
+        throw error;
+      })
+      .then(() => {
       // We use $exists and $set to avoid overwriting the field type if it
       // already exists. (it could have added inbetween the last query and the update)
-      return this.upsertSchema(
-        className,
-        { [fieldName]: { '$exists': false } },
-        { '$set' : { [fieldName]: parseFieldTypeToMongoFieldType(type) } }
-      );
-    });
+        return this.upsertSchema(
+          className,
+          { [fieldName]: { '$exists': false } },
+          { '$set' : { [fieldName]: parseFieldTypeToMongoFieldType(type) } }
+        );
+      });
   }
 }
 
