@@ -227,8 +227,8 @@ function matchesKeyConstraints(object, key, constraints) {
       const propertyExists = typeof object[key] !== 'undefined';
       const existenceIsRequired = constraints['$exists'];
       if (typeof constraints['$exists'] !== 'boolean') {
-          // The SDK will never submit a non-boolean for $exists, but if someone
-          // tries to submit a non-boolean for $exits outside the SDKs, just ignore it.
+        // The SDK will never submit a non-boolean for $exists, but if someone
+        // tries to submit a non-boolean for $exits outside the SDKs, just ignore it.
         break;
       }
       if ((!propertyExists && existenceIsRequired) || (propertyExists && !existenceIsRequired)) {
@@ -240,17 +240,17 @@ function matchesKeyConstraints(object, key, constraints) {
       if (typeof compareTo === 'object') {
         return compareTo.test(object[key]);
       }
-        // JS doesn't support perl-style escaping
+      // JS doesn't support perl-style escaping
       var expString = '';
       var escapeEnd = -2;
       var escapeStart = compareTo.indexOf('\\Q');
       while (escapeStart > -1) {
-          // Add the unescaped portion
+        // Add the unescaped portion
         expString += compareTo.substring(escapeEnd + 2, escapeStart);
         escapeEnd = compareTo.indexOf('\\E', escapeStart);
         if (escapeEnd > -1) {
           expString += compareTo.substring(escapeStart + 2, escapeEnd)
-              .replace(/\\\\\\\\E/g, '\\E').replace(/\W/g, '\\$&');
+            .replace(/\\\\\\\\E/g, '\\E').replace(/\W/g, '\\$&');
         }
 
         escapeStart = compareTo.indexOf('\\Q', escapeEnd);
@@ -270,22 +270,22 @@ function matchesKeyConstraints(object, key, constraints) {
       var northEast = compareTo.$box[1];
       if (southWest.latitude > northEast.latitude ||
             southWest.longitude > northEast.longitude) {
-          // Invalid box, crosses the date line
+        // Invalid box, crosses the date line
         return false;
       }
       return (
-          object[key].latitude > southWest.latitude &&
+        object[key].latitude > southWest.latitude &&
           object[key].latitude < northEast.latitude &&
           object[key].longitude > southWest.longitude &&
           object[key].longitude < northEast.longitude
       );
     case '$options':
-        // Not a query type, but a way to add options to $regex. Ignore and
-        // avoid the default
+      // Not a query type, but a way to add options to $regex. Ignore and
+      // avoid the default
       break;
     case '$maxDistance':
-        // Not a query type, but a way to add a cap to $nearSphere. Ignore and
-        // avoid the default
+      // Not a query type, but a way to add a cap to $nearSphere. Ignore and
+      // avoid the default
       break;
     case '$select':
       return false;
