@@ -32,7 +32,7 @@ export class FeaturesRouter extends PromiseRouter {
           immediatePush: req.config.hasPushSupport,
           scheduledPush: req.config.hasPushScheduledSupport,
           storedPushData: req.config.hasPushSupport,
-          pushAudiences: false,
+          pushAudiences: true,
         },
         schemas: {
           addField: true,
