@@ -5,15 +5,15 @@ export class PurgeRouter extends PromiseRouter {
 
   handlePurge(req) {
     return req.config.database.purgeCollection(req.params.className)
-    .then(() => {
-      var cacheAdapter = req.config.cacheController;
-      if (req.params.className == '_Session') {
-        cacheAdapter.user.clear();
-      } else if (req.params.className == '_Role') {
-        cacheAdapter.role.clear();
-      }
-      return {response: {}};
-    });
+      .then(() => {
+        var cacheAdapter = req.config.cacheController;
+        if (req.params.className == '_Session') {
+          cacheAdapter.user.clear();
+        } else if (req.params.className == '_Role') {
+          cacheAdapter.role.clear();
+        }
+        return {response: {}};
+      });
   }
 
   mountRoutes() {
