@@ -26,11 +26,11 @@ function makeBatchRoutingPathFunction(originalUrl, serverURL, publicServerURL) {
   let apiPrefix = originalUrl.slice(0, apiPrefixLength);
 
   const makeRoutablePath = function(requestPath) {
-      // The routablePath is the path minus the api prefix
+    // The routablePath is the path minus the api prefix
     if (requestPath.slice(0, apiPrefix.length) != apiPrefix) {
       throw new Parse.Error(
-          Parse.Error.INVALID_JSON,
-          'cannot route batch path ' + requestPath);
+        Parse.Error.INVALID_JSON,
+        'cannot route batch path ' + requestPath);
     }
     return path.posix.join('/', requestPath.slice(apiPrefix.length));
   }
@@ -39,13 +39,13 @@ function makeBatchRoutingPathFunction(originalUrl, serverURL, publicServerURL) {
         && (serverURL.path != publicServerURL.path)) {
     const localPath = serverURL.path;
     const publicPath = publicServerURL.path;
-      // Override the api prefix
+    // Override the api prefix
     apiPrefix = localPath;
     return function(requestPath) {
-        // Build the new path by removing the public path
-        // and joining with the local path
+      // Build the new path by removing the public path
+      // and joining with the local path
       const newPath = path.posix.join('/', localPath, '/' , requestPath.slice(publicPath.length));
-        // Use the method for local routing
+      // Use the method for local routing
       return makeRoutablePath(newPath);
     }
   }
@@ -58,7 +58,7 @@ function makeBatchRoutingPathFunction(originalUrl, serverURL, publicServerURL) {
 function handleBatch(router, req) {
   if (!Array.isArray(req.body.requests)) {
     throw new Parse.Error(Parse.Error.INVALID_JSON,
-                          'requests must be an array');
+      'requests must be an array');
   }
 
   // The batch paths are all from the root of our domain.
