@@ -133,7 +133,8 @@ export function getRequestObject(triggerType, auth, parseObject, originalParseOb
     triggerName: triggerType,
     object: parseObject,
     master: false,
-    log: config.loggerController
+    log: config.loggerController,
+    headers: config.headers,
   };
 
   if (originalParseObject) {
@@ -155,13 +156,17 @@ export function getRequestObject(triggerType, auth, parseObject, originalParseOb
   return request;
 }
 
-export function getRequestQueryObject(triggerType, auth, query, count, config) {
+export function getRequestQueryObject(triggerType, auth, query, count, config, isGet) {
+  isGet = !!isGet;
+
   var request = {
     triggerName: triggerType,
     query,
     master: false,
     count,
-    log: config.loggerController
+    log: config.loggerController,
+    isGet,
+    headers: config.headers,
   };
 
   if (!auth) {
@@ -286,7 +291,7 @@ export function maybeRunAfterFindTrigger(triggerType, auth, className, objects, 
   });
 }
 
-export function maybeRunQueryTrigger(triggerType, className, restWhere, restOptions, config, auth) {
+export function maybeRunQueryTrigger(triggerType, className, restWhere, restOptions, config, auth, isGet) {
   const trigger = getTrigger(className, triggerType, config.applicationId);
   if (!trigger) {
     return Promise.resolve({
@@ -312,7 +317,7 @@ export function maybeRunQueryTrigger(triggerType, className, restWhere, restOpti
     }
     count = !!restOptions.count;
   }
-  const requestObject = getRequestQueryObject(triggerType, auth, parseQuery, count, config);
+  const requestObject = getRequestQueryObject(triggerType, auth, parseQuery, count, config, isGet);
   return Promise.resolve().then(() => {
     return trigger(requestObject);
   }).then((result) => {
@@ -339,6 +344,18 @@ export function maybeRunQueryTrigger(triggerType, className, restWhere, restOpti
     if (jsonQuery.keys) {
       restOptions = restOptions || {};
       restOptions.keys = jsonQuery.keys;
+    }
+    if (requestObject.readPreference) {
+      restOptions = restOptions || {};
+      restOptions.readPreference = requestObject.readPreference;
+    }
+    if (requestObject.includeReadPreference) {
+      restOptions = restOptions || {};
+      restOptions.includeReadPreference = requestObject.includeReadPreference;
+    }
+    if (requestObject.subqueryReadPreference) {
+      restOptions = restOptions || {};
+      restOptions.subqueryReadPreference = requestObject.subqueryReadPreference;
     }
     return {
       restWhere,
@@ -368,11 +385,11 @@ export function maybeRunTrigger(triggerType, auth, parseObject, originalParseObj
     var request = getRequestObject(triggerType, auth, parseObject, originalParseObject, config);
     var response = getResponseObject(request, (object) => {
       logTriggerSuccessBeforeHook(
-          triggerType, parseObject.className, parseObject.toJSON(), object, auth);
+        triggerType, parseObject.className, parseObject.toJSON(), object, auth);
       resolve(object);
     }, (error) => {
       logTriggerErrorBeforeHook(
-          triggerType, parseObject.className, parseObject.toJSON(), auth, error);
+        triggerType, parseObject.className, parseObject.toJSON(), auth, error);
       reject(error);
     });
     // Force the current Parse app before the trigger
